% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFeatureFunction.R
\name{A5SS.length.feature}
\alias{A5SS.length.feature}
\title{get length of exons and introns associated with A5SS event}
\usage{
A5SS.length.feature(event)
}
\arguments{
\item{event}{a event id}
}
\value{
data.frame
}
\description{
get length of exons and introns associated with A5SS event
}
\keyword{internal}
