% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{addResourcePath}
\alias{addResourcePath}
\title{Resource Publishing}
\usage{
addResourcePath(prefix, directoryPath)
}
\arguments{
\item{prefix}{The URL prefix (without slashes). Valid characters are a-z,
A-Z, 0-9, hyphen, period, and underscore.
For example, a value of 'foo' means that any request paths that begin with
'/foo' will be mapped to the given directory.}

\item{directoryPath}{The directory that contains the static resources to be
served.}
}
\description{
Adds a directory of static resources to Shiny's web server, with the given
path prefix. Primarily intended for package authors to make supporting
JavaScript/CSS files available to their components.
}
\examples{
addResourcePath('datasets', system.file('data', package='datasets'))
}
\seealso{
\code{\link{singleton}}
}
