% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFeatureFunction.R
\name{AL.exon.seq.extraction}
\alias{AL.exon.seq.extraction}
\title{Extract exon sequence for AL event}
\usage{
AL.exon.seq.extraction(events.info, bs.genome, core)
}
\arguments{
\item{events.info}{data.frame AL event information}

\item{bs.genome}{\code{get(pkg)}, see \code{createBSgenome} function}

\item{core}{the number of threads}
}
\value{
list
}
\description{
Extract exon sequence for AL event
}
\keyword{internal}
