% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFeatureFunction.R
\name{AL.info.seperation}
\alias{AL.info.seperation}
\title{Split AL event information}
\usage{
AL.info.seperation(event)
}
\arguments{
\item{event}{a event id}
}
\value{
data.frame
}
\description{
Split AL event information
}
\keyword{internal}
