% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFeatureFunction.R
\name{AL.kmer.extraction}
\alias{AL.kmer.extraction}
\title{get kmer of exons and splice region associated with AL event}
\usage{
AL.kmer.extraction(splice.region, exon.region)
}
\arguments{
\item{splice.region}{data.frame with sequence of splice region, \code{AL.splice.seq.extraction}}

\item{exon.region}{data.frame with sequence of exon region, \code{AL.exon.seq.extraction}}
}
\value{
vector of sequence kmer
}
\description{
get kmer of exons and splice region associated with AL event
}
\keyword{internal}
