% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFeatureFunction.R
\name{AL.length.feature}
\alias{AL.length.feature}
\title{get length of exons and introns associated with AL event}
\usage{
AL.length.feature(event)
}
\arguments{
\item{event}{a event id}
}
\value{
data.frame
}
\description{
get length of exons and introns associated with AL event
}
\keyword{internal}
