% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{allWorklog2}
\alias{allWorklog2}
\title{R style wich returns the worklog of all the users.}
\usage{
allWorklog2(worklogsMatrix, start, end)
}
\description{
R style wich returns the worklog of all the users.
}

