% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFeatureFunction.R
\name{aPercentage}
\alias{aPercentage}
\title{Calculate alternative branch point selection
Specifically, for an alternative region in an event, we divide the region
into \bold{seperate} bins with equal length, and calculate the cumulative adenine ratio
for each bin as the adenine ratio features.}
\usage{
aPercentage(region, seperate = 100, core)
}
\arguments{
\item{region}{data.frame containing column event id and sequence}

\item{seperate}{the number of bins, default: 100}

\item{core}{the number of threads}
}
\value{
list
}
\description{
Calculate alternative branch point selection
Specifically, for an alternative region in an event, we divide the region
into \bold{seperate} bins with equal length, and calculate the cumulative adenine ratio
for each bin as the adenine ratio features.
}
\keyword{internal}
