% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-plot.R
\name{autoThemeSet}
\alias{autoThemeSet}
\alias{autoThemeGet}
\alias{autoThemeClear}
\title{Plot auto theming options}
\usage{
autoThemeSet(
  fg = "auto",
  bg = "auto",
  accent = "auto",
  font = NA,
  sequential = thematic::sequential_gradient(),
  qualitative = thematic::okabe_ito()
)

autoThemeGet()

autoThemeClear()
}
\arguments{
\item{fg}{a foreground color.}

\item{bg}{a background color.}

\item{accent}{a color for making certain graphical markers 'stand out'
(e.g., the fitted line color for \code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}}).
Can be 2 colors for lattice (stroke vs fill accent).}

\item{font}{a \code{font_spec()} object. If missing, font defaults are not altered.}

\item{sequential}{a color palette for graphical markers that encode
numeric values. Can be a vector of color codes or a
\code{\link[thematic:sequential_gradient]{sequential_gradient()}} object.}

\item{qualitative}{a color palette for graphical markers that encode
qualitative values (won't be used in ggplot2 when the number of data
levels exceeds the max allowed colors). Defaults to the Okabe-Ito colorscale.}
}
\description{
Plot auto theming options
}
