% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_example.R
\name{available_examples}
\alias{available_examples}
\title{Function to get a data frame of all available Shiny examples}
\usage{
available_examples(package = NULL)
}
\arguments{
\item{package}{A character string specifying the name of the package}
}
\value{
A data frame containing information
about all available Shiny examples
}
\description{
Function to get a data frame of all available Shiny examples
}
