% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-example.R
\name{availableExamples}
\alias{availableExamples}
\title{Function to get a data frame of all available Shiny examples}
\usage{
availableExamples(package = NULL)
}
\arguments{
\item{package}{A character string specifying the name of the package}
}
\value{
A data frame containing information
about all available Shiny examples
}
\description{
Function to get a data frame of all available Shiny examples
}
