% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended-task.R
\name{bind_button_to_task}
\alias{bind_button_to_task}
\title{Bind \code{input_task_button} to \code{ExtendedTask}}
\usage{
bind_button_to_task(
  task,
  task_button_id,
  priority = 1000,
  session = getDefaultReactiveDomain()
)
}
\description{
Sets up an \link{ExtendedTask} to relay its state to an existing
\code{\link[bslib:input_task_button]{bslib::input_task_button()}}, so the task button stays in its "busy" state
for as long as the extended task is running.

Note that \code{bind_button_to_task} does \emph{not} automatically cause button presses
to invoke the extended task; you still need to use \code{\link[=bindEvent]{bindEvent()}} (or
\code{\link[=observeEvent]{observeEvent()}}) to cause the button press to trigger
an invocation, as in the example below.
}
