% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-state.R
\name{bookmarkButton}
\alias{bookmarkButton}
\title{Create a button for bookmarking/sharing}
\usage{
bookmarkButton(inputId, label = tagList(icon("link", lib = "glyphicon"),
  "Share..."), ...)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button or link--usually a text label, but
you could also use any other HTML, like an image.}

\item{...}{Named attributes to be applied to the button or link.}
}
\description{
A \code{bookmarkButton} is a \code{\link{actionButton}} with a default label
that consists of a link icon and the text "Share...". It is meant to be used
for bookmarking state.
}
\seealso{
configureBookmarking
}

