% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hover-brush-options.R
\name{brushOptions}
\alias{brushOptions}
\title{Create an object representing brushing options}
\usage{
brushOptions(color = "#666", outline = "#000", opacity = 0.3,
  delay = 300, delayType = c("debounce", "throttle"), clip = TRUE,
  direction = c("xy", "x", "y"))
}
\arguments{
\item{color}{Fill color of the brush.}

\item{outline}{Outline color of the brush.}

\item{opacity}{Opacity of the brush}

\item{delay}{How long to delay (in milliseconds) when debouncing or
throttling, before sending the brush data to the server.}

\item{delayType}{The type of algorithm for limiting the number of brush
events. Use \code{"throttle"} to limit the number of brush events to one
every \code{delay} milliseconds. Use \code{"debounce"} to suspend events
while the cursor is moving, and wait until the cursor has been at rest for
\code{delay} milliseconds before sending an event.}

\item{clip}{Should the brush area be clipped to the plotting area? If FALSE,
then the user will be able to brush outside the plotting area, as long as
it is still inside the image.}

\item{direction}{The direction for brushing. If \code{"xy"}, the brush can be
drawn and moved in both x and y directions. If \code{"x"}, or \code{"y"},
the brush wil work horizontally or vertically.}
}
\description{
This generates an object representing brushing options, to be passed as the
\code{brushOpts} argument of \code{\link{imageOutput}} or
\code{\link{plotOutput}}.
}

