% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{buildDFissues}
\alias{buildDFissues}
\title{Creates a data frame with basic data of the sprint's users.}
\usage{
buildDFissues(issues)
}
\arguments{
\item{issues}{A list with the json of the issues in the sprint}
}
\value{
Data frame with columns "Issue key", "assignee key", "assignee name", "status", "type", "original estimate (h)", "time spent (h)", "difference (h)", "story points", "story points filled", "parent", "time judgement"
}
\description{
Creates a data frame with basic data of the sprint's users.
}

