% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modules.R
\name{callModule}
\alias{callModule}
\title{Invoke a module}
\usage{
callModule(module, id, ..., session = getDefaultReactiveDomain())
}
\arguments{
\item{module}{A function or path}

\item{id}{An ID string}

\item{...}{Additional parameters to pass to module function}

\item{session}{Session from which to make a child scope}
}
\description{
Invoke a module
}

