% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{cardPanel}
\alias{cardPanel}
\title{Create a card panel}
\usage{
cardPanel(...)
}
\arguments{
\item{...}{UI elements to include inside the panel.}
}
\value{
The newly created panel.
}
\description{
Creates a panel with a slightly inset border and grey background. Equivalent
to Bootstrap's \code{card} CSS class. Bootstrap v4 replaces \code{well} CSS
class with \code{card}s class.
}
