% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{closeShinySessions}
\alias{closeShinySessions}
\title{Close all opened websocket connections}
\usage{
closeShinySessions()
}
\description{
This function iterate on available user session and trigger \code{wsClosed}
to ends them.
}
