\name{columnLayout}
\alias{columnLayout}
\title{Layout a set of columns}
\usage{
  columnLayout(...)
}
\arguments{
  \item{...}{Columns to include within the layout}
}
\description{
  Layout a set of columns created using the
  \code{\link{column}} function. The widths of the columns
  should total no more than 12 units.
}

