\name{containerPanel}
\alias{containerPanel}
\title{Create a container panel}
\usage{
  containerPanel(...)
}
\arguments{
  \item{...}{Elements to include within the container}
}
\description{
  Create a container that includes one or more rows of
  content (each element passed to the container will appear
  on it's own line in the UI)
}

