% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createWebDependency}
\alias{createWebDependency}
\title{Create a web dependency}
\usage{
createWebDependency(dependency)
}
\arguments{
\item{dependency}{A single HTML dependency object, created using
  \code{\link{htmlDependency}}. If the \code{src} value is named, then
  \code{href} and/or \code{file} names must be present.}
}
\value{
A single HTML dependency object that has an \code{href}-named element
  in its \code{src}.
}
\description{
Ensure that a file-based HTML dependency (from the htmltools package) can be
served over Shiny's HTTP server. This function works by using
\code{\link{addResourcePath}} to map the HTML dependency's directory to a
URL.
}

