% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-date.R
\name{dateInput}
\alias{dateInput}
\title{Create date input}
\usage{
dateInput(inputId, label, value = NULL, min = NULL, max = NULL,
  format = "yyyy-mm-dd", startview = "month", weekstart = 0,
  language = "en", autoclose = FALSE, multidate = FALSE,
  multidateSeparator = ",", orientation = "auto", title = NULL,
  todayBtn = FALSE, todayHighlight = FALSE, toggleActive = FALSE,
  calendarWeeks = FALSE, clearBtn = FALSE, datesDisabled = NULL,
  daysOfWeekDisabled = NULL, daysOfWeekHighlighted = NULL,
  immediateUpdates = FALSE, keyboardNavigation = TRUE, width = NULL)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{The starting date. Either a Date object, or a string in
\code{yyyy-mm-dd} format. If NULL (the default), will use the current date
in the client's time zone.}

\item{min}{The minimum allowed date. Either a Date object, or a string in
\code{yyyy-mm-dd} format.}

\item{max}{The maximum allowed date. Either a Date object, or a string in
\code{yyyy-mm-dd} format.}

\item{format}{The format of the date to display in the browser. Defaults to
\code{"yyyy-mm-dd"}.}

\item{startview}{The date range shown when the input object is first clicked.
Can be "month" (the default), "year", or "decade".}

\item{weekstart}{Which day is the start of the week. Should be an integer
from 0 (Sunday) to 6 (Saturday).}

\item{language}{The language used for month and day names. Default is "en".
Other valid values include "ar", "az", "bg", "bs", "ca", "cs", "cy", "da",
"de", "el", "en-AU", "en-GB", "eo", "es", "et", "eu", "fa", "fi", "fo",
"fr-CH", "fr", "gl", "he", "hr", "hu", "hy", "id", "is", "it-CH", "it",
"ja", "ka", "kh", "kk", "ko", "kr", "lt", "lv", "me", "mk", "mn", "ms",
"nb", "nl-BE", "nl", "no", "pl", "pt-BR", "pt", "ro", "rs-latin", "rs",
"ru", "sk", "sl", "sq", "sr-latin", "sr", "sv", "sw", "th", "tr", "uk",
"vi", "zh-CN", and "zh-TW".}

\item{autoclose}{Close immediately when a date is selected (\code{TRUE} or
\code{FALSE}).}

\item{multidate}{Enable multidate. Either \code{TRUE} or \code{FALSE} or an
integer stating how  many dates can be selected, dropping the oldest dates
from the list when the number is exceeded. The input's value is set to a
string generated by joining the dates, formatted, with
\code{multidateSeparator}.}

\item{multidateSeparator}{The string that will appear between dates when
generating the multidate input's value.}

\item{orientation}{Location of the popup's anchor. A space-separated string
consisting of one or two of \code{left} or \code{right}, \code{top}
or \code{bottom}, and \code{auto}.}

\item{title}{Title on the top of the date picker.}

\item{todayBtn}{Display a "Today" button. If \code{TRUE} it will move the
current date into view. If \code{"linked"} the current date will also be
selected.}

\item{todayHighlight}{Highlight the current date.}

\item{toggleActive}{If \code{TRUE}, selecting the currently active date in
the datepicker will unset the respective date. Always \code{TRUE} when
multidate is being used.}

\item{calendarWeeks}{Show week numbers to the left of week rows (\code{TRUE}
or \code{FALSE}).}

\item{clearBtn}{Display a "Clear" button (\code{TRUE} or \code{FALSE}).}

\item{datesDisabled}{Disabled dates. Either a Date object, a string
in \code{yyyy-mm-dd} format or a vector.}

\item{daysOfWeekDisabled}{Disabled days of week. Should be an integer from 0
(Sunday) to 6 (Saturday) or a vector.}

\item{daysOfWeekHighlighted}{Highlighted days of week. Should be an integer
from 0 (Sunday) to 6 (Saturday) or a vector.}

\item{immediateUpdates}{Update the input value immediately when selecting a
year or month (\code{TRUE} or \code{FALSE}).}

\item{keyboardNavigation}{Allow date navigation by arrow keys (\code{TRUE} or
\code{FALSE}).}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link{validateCssUnit}}.}
}
\description{
Creates a text input which, when clicked on, brings up a calendar that
the user can click on to select dates.
}
\details{
The date \code{format} string specifies how the date will be displayed in
the browser. It allows the following values:

\itemize{
  \item \code{yy} Year without century (12)
  \item \code{yyyy} Year with century (2012)
  \item \code{mm} Month number, with leading zero (01-12)
  \item \code{m} Month number, without leading zero (1-12)
  \item \code{M} Abbreviated month name
  \item \code{MM} Full month name
  \item \code{dd} Day of month with leading zero
  \item \code{d} Day of month without leading zero
  \item \code{D} Abbreviated weekday name
  \item \code{DD} Full weekday name
}
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {

ui <- fluidPage(
  dateInput("date1", "Date:", value = "2012-02-29"),

  # Default value is the date in client's time zone
  dateInput("date2", "Date:"),

  # value is always yyyy-mm-dd, even if the display format is different
  dateInput("date3", "Date:", value = "2012-02-29", format = "mm/dd/yy"),

  # Pass in a Date object
  dateInput("date4", "Date:", value = Sys.Date()-10),

  # Use different language and different first day of week
  dateInput("date5", "Date:",
          language = "ru",
          weekstart = 1),

  # Start with decade view instead of default month view
  dateInput("date6", "Date:",
            startview = "decade")
)

shinyApp(ui, server = function(input, output) { })
}
}
\seealso{
\code{\link{dateRangeInput}}, \code{\link{updateDateInput}}

Other input elements: \code{\link{actionButton}},
  \code{\link{checkboxGroupInput}},
  \code{\link{checkboxInput}},
  \code{\link{dateRangeInput}}, \code{\link{fileInput}},
  \code{\link{numericInput}}, \code{\link{passwordInput}},
  \code{\link{radioButtons}}, \code{\link{selectInput}},
  \code{\link{sliderInput}}, \code{\link{submitButton}},
  \code{\link{textAreaInput}}, \code{\link{textInput}}
}
