% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-interact-opts.R
\name{dblclickOpts}
\alias{dblclickOpts}
\title{Create an object representing double-click options}
\usage{
dblclickOpts(id, clip = TRUE, delay = 400)
}
\arguments{
\item{id}{Input value name. For example, if the value is
\code{"plot_dblclick"}, then the click coordinates will be available as
\code{input$plot_dblclick}.}

\item{clip}{Should the click area be clipped to the plotting area? If FALSE,
then the server will receive double-click events even when the mouse is
outside the plotting area, as long as it is still inside the image.}

\item{delay}{Maximum delay (in ms) between a pair clicks for them to be
counted as a double-click.}
}
\description{
This generates an object representing dobule-click options, to be passed as
the \code{dblclick} argument of \code{\link[=imageOutput]{imageOutput()}} or
\code{\link[=plotOutput]{plotOutput()}}.
}
