% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{Estimation}
\alias{Estimation}
\title{Combining different imputation strategies}
\usage{
Estimation(
  paras,
  rds_imputed_file,
  output_path = NULL,
  cell_similarity = NULL,
  dyk_cell = NULL,
  rc = NULL,
  psi = NULL,
  expr = NULL,
  event = NULL
)
}
\arguments{
\item{paras}{list fromJSON(paras_file)}

\item{rds_imputed_file}{path to the list of three imputation strategies results
named by the strategies}

\item{output_path}{path to save the final imputation of PSI values}

\item{cell_similarity}{a list of cell similarity named by the data
type (EXP_RBP, RC, and PSI) or path to a rds file
default: imputation/cell_similarity/cell.similars.rds}

\item{dyk_cell}{a list of Neighbor number named by the data
type (EXP_RBP, RC, and PSI) or path to a rds file
default: imputation/cell_similarity/dyk.cell.rds}

\item{rc}{matrix of normalized splicing events associated read counts or path
to a rds file, default: rds_processed/rc.rds}

\item{psi}{matrix of psi value or path to a rds file,
default: rds_processed/psi.rds}

\item{expr}{matrix of gene expression or path to a rds file, TPM for full-length protocols
and normalized UMI counts for droplet-based protocols
default: rds_processed/expr.rds}

\item{event}{data.frame of events information or path to a rds file,
default: rds_processed/event.rds}
}
\value{
path to save the final imputation of PSI values
The names of the lists represent different cell similarities
}
\description{
strategy1 for pairs in ND,
strategy2 for pairs in BD and TD+Info, and strategy3 for pairs in T0-Info.
For each event-cell pair, the different imputation results are combined with
the probabilities of different groups as the coefficients.
}
