% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactives.R
\name{format.reactivevalues}
\alias{format.reactivevalues}
\alias{print.reactivevalues}
\title{Format or print reactivevalues object}
\usage{
\method{format}{reactivevalues}(x, all.names = TRUE, ...)
}
\arguments{
\item{x}{A reactivevalues object.}

\item{all.names}{If \code{TRUE} (the default), include objects with a
leading dot. If FALSE, don't include those objects.}

\item{...}{Not used.}
}
\description{
\code{format.reactivevalues} creates a textual representation of a
reactivevalues object, whereas \code{print.reactivevalues} prints it.
}
\details{
\code{format.reactivevalues} calls \link{reactiveValuesToList} internally.
Thereby its calling context will take a dependency on all the values
of \code{x}. If you want to avoid it, wrap the call with \code{isolate()}.\cr
\code{print.reactivevalues} does not take a dependency on the reactive objects.
}
