% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{getClassifierFeature}
\alias{getClassifierFeature}
\title{Calculate features for classifer}
\usage{
getClassifierFeature(
  cell_similarity,
  psi,
  rc,
  expr,
  event,
  k,
  rc_imputed,
  psi_imputed
)
}
\arguments{
\item{cell_similarity}{matrix of cell similarity}

\item{psi}{matrix of psi value}

\item{rc}{matrix of normalized splicing events associated read counts}

\item{expr}{matrix of gene expression (TPM for full-length protocols
and normalized UMI counts for droplet-based protocols)}

\item{event}{data.frame of events information}

\item{k}{the number of neighbor cells}

\item{rc_imputed}{matrix of psi value calculated using reads after
imputation with information from similar cells  (Strategy 2)}

\item{psi_imputed}{matrix of imputed psi with information from similar cells (Strategy 1)}
}
\value{
a data.frame of features
}
\description{
Calculate features for classifer
}
