\name{getCurrentContext}
\alias{getCurrentContext}
\title{Retrieves the current reactive context, or errors if there is no reactive
context active at the moment.}
\usage{
  getCurrentContext()
}
\description{
  Retrieves the current reactive context, or errors if
  there is no reactive context active at the moment.
}

