% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{getCurrentOutputInfo}
\alias{getCurrentOutputInfo}
\title{Get information about the output that is currently being executed.}
\usage{
getCurrentOutputInfo(session = getDefaultReactiveDomain())
}
\arguments{
\item{session}{The current Shiny session.}
}
\value{
A list with information about the current output, including the
\code{name} of the output. If no output is currently being executed, this will
return \code{NULL}.
}
\description{
Get information about the output that is currently being executed.
}
