% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{getIndividualIssuesJson}
\alias{getIndividualIssuesJson}
\title{Creates a matrix with the issues from a list.}
\usage{
getIndividualIssuesJson(issueKeys)
}
\arguments{
\item{issuesKeys}{A list of factors with the keys of the issues}
}
\value{
matrix with 7 columns "issue", "assignee", "expand", "id", "self", "field".  The column "assignee" is the key. The column "fields" is the original json from the issue.
}
\description{
Creates a matrix with the issues from a list.
}

