% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getProvidedHtmlDependencies}
\alias{getProvidedHtmlDependencies}
\title{Return HTML dependencies provided by Shiny}
\usage{
getProvidedHtmlDependencies()
}
\value{
A list of objects of type \code{html_dependency}, one per dependency
}
\description{
By default, Shiny supplies some framework scripts when it renders a page.
\code{getProvidedHtmlDependencies} returns a list of those provided objects.
}

