% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectEvent.R
\name{getRIid}
\alias{getRIid}
\title{Generate RI event id}
\usage{
getRIid(
  file,
  base = 5,
  outfile,
  remove.chr = F,
  junctionReads,
  ExonToIntronReads,
  gtf
)
}
\arguments{
\item{file}{directory to IRFinder output}

\item{base}{default=5}

\item{outfile}{directory to output file RI.txt}

\item{remove.chr}{if TRUE,remove the "chr" character of event id}

\item{junctionReads}{the minimum total number of reads for any junction}

\item{ExonToIntronReads}{the minimum total number of reads that overlap the 3'/5' flanking exon and the intron}

\item{gtf}{the gene annotation in gtf format,used to get exon position}
}
\value{
The number of RI events
}
\description{
Generate RI event id
}
