% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectEvent.R
\name{getSEid}
\alias{getSEid}
\title{Generate SE event id}
\usage{
getSEid(file, outfile, remove.chr = F, junctionReads)
}
\arguments{
\item{file}{directory to rMats output}

\item{outfile}{directory to output file SE.txt}

\item{remove.chr}{if TRUE,remove the "chr" character of event id}

\item{junctionReads}{the minimum total number of reads for any junction}
}
\value{
The number of SE events
}
\description{
Generate SE event id
}
