\name{headerPanel}
\alias{headerPanel}
\title{Create a header panel}
\usage{
  headerPanel(title)
}
\arguments{
  \item{title}{An application title to display}
}
\value{
  A headerPanel that can be passed to
  \link{pageWithSidebar}
}
\description{
  Create a header panel containing an application title.
}
\examples{
headerPanel("Hello Shiny!")
}

