\name{horizontalLayout}
\alias{horizontalLayout}
\alias{pullLeft}
\alias{pullRight}
\title{Layout UI elements horizontally}
\usage{
  horizontalLayout(...)

  pullLeft(...)

  pullRight(...)
}
\arguments{
  \item{...}{Elements to include within the container}

  \item{...}{Element or list of elements to pull left or
  right}
}
\description{
  Create a container that includes several elements laid
  out side-by-side.

  Pull an element to the left or right side of a
  \code{\link{horizontalLayout}}.
}
\details{
  To force elements to the left or right of the container
  you can use the \code{\link{pullLeft}} and
  \code{\link{pullRight}} functions.
}

