% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hover-brush-options.R
\name{hoverOptions}
\alias{hoverOptions}
\title{Create an object representing hover options}
\usage{
hoverOptions(delay = 300, delayType = c("debounce", "throttle"),
  clip = TRUE)
}
\arguments{
\item{delay}{How long to delay (in milliseconds) when debouncing or
throttling, before sending the mouse location to the server.}

\item{delayType}{The type of algorithm for limiting the number of hover
events. Use \code{"throttle"} to limit the number of hover events to one
every \code{delay} milliseconds. Use \code{"debounce"} to suspend events
while the cursor is moving, and wait until the cursor has been at rest for
\code{delay} milliseconds before sending an event.}

\item{clip}{Should the hover area be clipped to the plotting area? If FALSE,
then the server will receive hover events even when the mouse is outside
the plotting area, as long as it is still inside the image.}
}
\description{
This generates an object representing hovering options, to be passed as the
\code{hoverOpts} argument of \code{\link{imageOutput}} or
\code{\link{plotOutput}}.
}

