% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{htmlTemplate}
\alias{htmlTemplate}
\title{Process an HTML template}
\usage{
htmlTemplate(filename, ..., document_ = "auto")
}
\arguments{
\item{filename}{Path to an HTML template file.}

\item{...}{Variable values to use when processing the template.}

\item{document_}{Is this template a complete HTML document (\code{TRUE}), or
a fragment of HTML that is to be inserted into an HTML document
(\code{FALSE})? With \code{"auto"} (the default), auto-detect by searching
for the string \code{"<HTML>"} within the template.}
}
\description{
Process an HTML template
}

