% Generated by roxygen2 (4.0.1): do not edit by hand
\name{inputPanel}
\alias{inputPanel}
\title{Input panel}
\usage{
inputPanel(...)
}
\arguments{
\item{...}{Input controls or other HTML elements.}
}
\description{
A \code{\link{flowLayout}} with a grey border and light grey background,
suitable for wrapping inputs.
}

