% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-ui.R
\name{insertUI}
\alias{insertUI}
\title{Insert UI objects}
\usage{
insertUI(selector, ui, where = c("end", "begin", "before", "after",
  "contents"), immediate = FALSE, session = getDefaultReactiveDomain())
}
\description{
Insert a UI object into the web page.
}

