% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactives.R
\name{invalidateReactiveValue}
\alias{invalidateReactiveValue}
\title{Invalidate a reactive value}
\usage{
invalidateReactiveValue(x, name)
}
\arguments{
\item{x}{A \code{\link{reactiveValues}} object (like \code{input}).}

\item{name}{The name of a value in the \code{\link{reactiveValues}} object.}
}
\description{
This invalidates a reactive value. If the value is accessed while invalid, a
"silent" exception is raised and the operation is stopped. This is the same
thing that happens if \code{req(FALSE)} is called. The value is
un-invalidated (accessing it will no longer raise an exception) when the
current reactive domain is flushed; in a Shiny application, this occurs after
all of the observers are executed.
}
\seealso{
\code{\link{req}}
}

