% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{isRunning}
\alias{isRunning}
\title{Check whether the Shiny app is running}
\usage{
isRunning()
}
\value{
\code{TRUE} if the function is called from within a running Shiny
  app. Otherwise, \code{FALSE}.
}
\description{
This function tests whether it is being called from within a running Shiny
app.
}
