% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{issueListWorklog}
\alias{issueListWorklog}
\title{Gets the worklogs of a issue list keys.}
\usage{
issueListWorklog(assignee, worklogs, start, end)
}
\arguments{
\item{worklogs}{A matrix with all the worklogs of the sprint}

\item{issueList}{A vector with keys of lists}
}
\description{
Gets the worklogs of a issue list keys.
}
\examples{
issueListWorklog(vector("PGR-64", "CADMF-219", "BDATA-19"), wl)
}

