% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{data}
\name{jiraProject}
\alias{jiraProject}
\title{jiraProject}
\format{An object of class \code{character} of length 1.}
\usage{
jiraProject
}
\description{
Used only to get the users of the Jira
}
\keyword{datasets}

