% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-utils.R
\name{key_missing}
\alias{key_missing}
\alias{is.key_missing}
\title{A Key Missing object}
\usage{
key_missing()

is.key_missing(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
A \code{key_missing} object represents a cache miss.
}
\seealso{
\code{\link[=diskCache]{diskCache()}}, \code{\link[=memoryCache]{memoryCache()}}.
}
