% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{knn_cv}
\alias{knn_cv}
\title{Calculate the coefficient of variation for psi of neighbor cells}
\usage{
knn_cv(psi, cs)
}
\arguments{
\item{psi}{matrix, psi associated with splicing events}

\item{cs}{matrix, cell similarity}
}
\value{
matrix
}
\description{
associated reads name recorded in event
}
\keyword{internal}
