\name{launchApp}
\alias{launchApp}
\title{Run an application in a new process and open it in a web browser}
\usage{
  launchApp(appDir, port = 8101L)
}
\arguments{
  \item{appDir}{Application directory}

  \item{port}{The TCP port that the application should
  listen on}
}
\value{
  The URL used to access the application.
}
\description{
  Run an application in a new process and open it in a web
  browser
}

