% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactives.R
\name{maskReactiveContext}
\alias{maskReactiveContext}
\title{Evaluate an expression without a reactive context}
\usage{
maskReactiveContext(expr)
}
\arguments{
\item{expr}{An expression to evaluate.}
}
\value{
The value of \code{expr}.
}
\description{
Temporarily blocks the current reactive context and evaluates the given
expression. Any attempt to directly access reactive values or expressions in
\code{expr} will give the same results as doing it at the top-level (by
default, an error).
}
\seealso{
\code{\link[=isolate]{isolate()}}
}
