% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeSplicingValue.R
\name{mgergePSI}
\alias{mgergePSI}
\title{Merge PSI}
\usage{
mgergePSI(splicing_value_path, output_path)
}
\arguments{
\item{splicing_value_path}{The path to *_psi.rds
* represents different types of splicing events}

\item{output_path}{The path to save merged PSI}
}
\value{
Merged PSI path
}
\description{
Merge PSI for different types of splicing events
in splicing_value_path, and save merged PSI to output_path
}
