% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{migrateLegacyShinytest}
\alias{migrateLegacyShinytest}
\title{Migrate legacy \pkg{shinytest} files to new test directory structure}
\usage{
migrateLegacyShinytest(appdir, prompt = interactive())
}
\arguments{
\item{appdir}{A directory containing a Shiny application.}

\item{prompt}{If \code{TRUE}, ask for confirmation when moving files.}
}
\description{
This function migrates the old-style directory structure used by
\pkg{shinytest} to (versions 1.3.1 and below) new test directory structure
used in Shiny 1.5.0 and above.
}
\details{
In Shiny 1.5.0, the \code{\link[=runTests]{runTests()}} function was added, and it will run test
scripts tests/ subdirectory of the application. The directory structure will
look something like this:\preformatted{appdir/
 |- R
 |- tests
     |- shinytest.R
     |- shinytest
     |   `- mytest.R
     |- testthat.R
     `- testthat
         `- test-script.R
}

This allows for tests using the \pkg{shinytest} package as well as other
testing tools, such as the \code{\link[=testServer]{testServer()}} function, which can be used for
testing module and server logic, and for unit tests of functions in an R/
subdirectory.

With the \pkg{shinytest} package, in versions 1.3.0 and below, the tests/
subdirectory of the application was used specifically for \pkg{shinytest},
and could not be used for other types of tests. So the directory structure
would look like this:\preformatted{appdir/
 `- tests
     `- mytest.R
}

In \pkg{shinytest} 1.4.0 and above, it defaults to the new directory
structure.
}
