% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_consulta_orc.R
\name{mod_consulta_orc_ui}
\alias{mod_consulta_orc_ui}
\alias{mod_consulta_orc_server}
\title{mod_consulta_orc_ui and mod_consulta_orc_server}
\usage{
mod_consulta_orc_ui(id)

mod_consulta_orc_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
