% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modalButton}
\alias{modalButton}
\title{Create a button for a modal dialog}
\usage{
modalButton(label, icon = NULL)
}
\arguments{
\item{label}{The contents of the button or link--usually a text label, but
you could also use any other HTML, like an image.}

\item{icon}{An optional \code{\link[=icon]{icon()}} to appear on the button.}
}
\description{
When clicked, a \code{modalButton} will dismiss the modal dialog.
}
\seealso{
\code{\link[=modalDialog]{modalDialog()}} for examples.
}
