% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modules.R
\name{moduleUI}
\alias{moduleUI}
\title{Invoke a module's UI}
\usage{
moduleUI(module, id, ...)
}
\arguments{
\item{module}{A function or path}

\item{id}{An ID string}

\item{...}{Additional parameters to pass to module function}
}
\value{
A tag, or a list of tags, or a \code{tagList} of tags
}
\description{
Invoke a module's UI
}

