% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFeatureFunction.R
\name{MXE.kmer.extraction}
\alias{MXE.kmer.extraction}
\title{get kmer of exons and splice region associated with MXE event}
\usage{
MXE.kmer.extraction(splice.region, exon.region)
}
\arguments{
\item{splice.region}{data.frame with sequence of splice region, \code{MXE.splice.seq.extraction}}

\item{exon.region}{data.frame with sequence of exon region, \code{MXE.exon.seq.extraction}}
}
\value{
vector of sequence kmer
}
\description{
get kmer of exons and splice region associated with MXE event
}
\keyword{internal}
