% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{navlistPanel}
\alias{navlistPanel}
\title{Create a navigation list panel}
\usage{
navlistPanel(
  ...,
  id = NULL,
  selected = NULL,
  well = TRUE,
  fluid = TRUE,
  widths = c(4, 8)
)
}
\arguments{
\item{...}{\code{\link[=tabPanel]{tabPanel()}} elements to include in the navbar.
Plain strings will be converted to headers.}

\item{id}{If provided, you can use \verb{input$}\emph{\code{id}} in your
server logic to determine which of the current tabs is active. The value
will correspond to the \code{value} argument that is passed to
\code{\link[=tabPanel]{tabPanel()}}.}

\item{selected}{The \code{value} (or, if none was supplied, the \code{title})
of the tab that should be selected by default. If \code{NULL}, the first
tab will be selected.}

\item{well}{\code{TRUE} to place a well (gray rounded rectangle) around the
navigation list.}

\item{fluid}{\code{TRUE} to use fluid layout; \code{FALSE} to use fixed
layout.}

\item{widths}{Column widths of the navigation list and tabset content areas
respectively.}
}
\description{
Create a navigation list panel that provides a list of links on the left
which navigate to a set of tabPanels displayed to the right.
}
\details{

}
\examples{
fluidPage(

  titlePanel("Application Title"),

  navlistPanel(
    "Header",
    tabPanel("First"),
    tabPanel("Second"),
    tabPanel("Third")
  )
)
}
\seealso{
\code{\link[=updateTabsetPanel]{updateTabsetPanel()}}, \code{\link[=insertTab]{insertTab()}}, \code{\link[=showTab]{showTab()}}

Other tab layouts: 
\code{\link{navbarPage}()},
\code{\link{tabPanel}()},
\code{\link{tabsetPanel}()}
}
\concept{tab layouts}
