% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-snapshot.R
\name{onTestSnapshot}
\alias{onTestSnapshot}
\title{Register expressions that will be evaluated when a test snapshot occurs}
\usage{
onTestSnapshot(..., quoted_ = FALSE, env_ = parent.frame(),
  session_ = getDefaultReactiveDomain())
}
\arguments{
\item{...}{Named arguments that are quoted or unquoted expressions that will
be captured and evaluated when API endpoint is visited.}

\item{quoted_}{Are the expression quoted? Default is \code{FALSE}.}

\item{env_}{The environment in which the expression should be evaluated.}

\item{session_}{A Shiny session object.}
}
\description{
This function registers expressions that will be called when a test snapshot
occurs. These test snapshots are available at an API endpoint URL.
}
\details{
This function only has an effect if the global option \code{shiny.testmode} is
set to \code{TRUE}.
}

