% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{packageLoading}
\alias{packageLoading}
\title{Prepare packages and general functions

Packages installing and loading and the functions
general to the application
Package installing and loading.}
\usage{
packageLoading()
}

