% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-state.R
\name{persistStateQueryString}
\alias{encodeStateQueryString}
\alias{persistStateQueryString}
\title{Persist (save) or encode state of Shiny session}
\usage{
persistStateQueryString(input, exclude = NULL, values = NULL)

encodeStateQueryString(input, exclude = NULL, values = NULL)
}
\arguments{
\item{input}{The session's input object.}

\item{exclude}{A character vector of input names that should not be
bookmarked.}

\item{values}{Any additional values that should be persisted or encoded. This
must be either NULL or a list.}
}
\description{
Shiny applications can have their state \emph{encoded} in a URL or
\emph{persisted}. If the state is encoded, all of the input values are stored in
the URL. If the state is persisted, the input values and any uploaded files are
stored on disk.
}

