% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageutils.R
\name{plotPNG}
\alias{plotPNG}
\title{Run a plotting function and save the output as a PNG}
\usage{
plotPNG(
  func,
  filename = tempfile(fileext = ".png"),
  width = 400,
  height = 400,
  res = 72,
  ...,
  device = getShinyOption("shiny.device", shinyDevice())
)
}
\arguments{
\item{func}{A function that generates a plot.}

\item{filename}{The name of the output file. Defaults to a temp file with
extension \code{.png}.}

\item{width}{Width in pixels.}

\item{height}{Height in pixels.}

\item{res}{Resolution in pixels per inch. This value is passed to
\code{device}. Note that this affects the resolution of PNG rendering in
R; it won't change the actual ppi of the browser.}

\item{...}{Arguments to be passed through to \code{device}.
These can be used to set the width, height, background color, etc.}

\item{device}{A function that opens a png graphics device (e.g., \code{\link[=png]{png()}}).}
}
\description{
This function returns the name of the PNG file that it generates. In
essence, it calls \code{device()}, then \code{func()}, then \code{dev.off()}.
So \code{func} must be a function that will generate a plot when used this
way.
}
\details{
The default \code{device} (i.e., \code{shinyDevice()}) is determined as follows:
\itemize{
\item If \code{\link[=quartz]{quartz()}} is operational, then \code{\link[=png]{png()}}.
\item If installed and \code{options(shiny.usecairo=TRUE)} (the default), then \code{\link[Cairo:CairoPNG]{Cairo::CairoPNG()}}.
\item Otherwise, \code{\link[=png]{png()}}.
}
}
