% Generated by roxygen2 (4.0.0): do not edit by hand
\name{print.shiny.appobj}
\alias{print.shiny.appobj}
\title{Run a Shiny app object}
\usage{
\method{print}{shiny.appobj}(x, ...)
}
\arguments{
\item{x}{A Shiny app, as returned from \code{\link{shinyAppObj}} or
  \code{\link{shinyAppDir}}.}
}
\description{
Run a Shiny app object
}

