\name{pullLeft}
\alias{pullLeft}
\alias{pullRight}
\title{Pull elements left or right}
\usage{
  pullLeft(element)

  pullRight(element)
}
\arguments{
  \item{element}{Element to pull left or right}
}
\description{
  Pull an element to the left or right side of a
  \code{\link{horizontalLayout}}.
}

