% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busy-indicators.R
\name{pulseOptions}
\alias{pulseOptions}
\title{Customize the pulsing busy indicator.}
\usage{
pulseOptions(color = NULL, height = NULL, speed = NULL)
}
\arguments{
\item{color}{The color of the pulsing banner. This can be any valid CSS
color. Defaults to the app's "primary" color (if Bootstrap is on the page)
or light-blue if not.}

\item{height}{The height of the pulsing banner. This can be any valid CSS
size. Defaults to "3.5px".}
}
\description{
Include the result of this function in the app's UI to customize the pulsing
banner.
}
\seealso{
\code{\link[=useBusyIndicators]{useBusyIndicators()}} \code{\link[=spinnerOptions]{spinnerOptions()}}
}
