% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{pushState}
\alias{pushState}
\title{Update URL in browser's location bar}
\usage{
pushState(state, title, url, session = getDefaultReactiveDomain())
}
\arguments{
\item{state}{State to save on the history stack.}

\item{title}{Not currently in use.}

\item{url}{New url to display on the window.}

\item{session}{A Shiny session object.}
}
\description{
This function updates the client browser's query string in the location bar.
It typically is called from an observer. Note that this will not work in
Internet Explorer 9 and below.
}
\seealso{
\code{\link{enableBookmarking}} for examples.
}

