% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quoToFunction}
\alias{quoToFunction}
\title{Convert a quosure to a function for a Shiny render function}
\usage{
quoToFunction(q, label, ..stacktraceon = FALSE)
}
\arguments{
\item{q}{A quosure.}

\item{label}{A label for the object to be shown in the debugger. Defaults to
the name of the calling function.}

\item{..stacktraceon}{Advanced use only. For stack
manipulation purposes; see \code{\link[=stacktrace]{stacktrace()}}.}
}
\description{
Superceded by \code{\link[=exprToFunction]{exprToFunction()}}: use that function instead for a simpler
and more robust implementation of custom render functions.
}
\seealso{
\code{\link[=createRenderFunction]{createRenderFunction()}} for example usage.
}
\keyword{internal}
