% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{randomPort}
\alias{randomPort}
\title{Find an open TCP port}
\usage{
randomPort(min = 3000L, max = 8000L, host = getOption("shiny.host",
  "127.0.0.1"), n = 20, cache = TRUE)
}
\arguments{
\item{min}{Minimum port number.}

\item{max}{Maximum port number.}

\item{host}{see \code{\link[httpuv:randomPort]{httpuv::randomPort()}}.}

\item{n}{Number of ports to try before giving up.}

\item{cache}{if available, reuse the last random port.}
}
\description{
Finds a random available TCP port for listening on.
The default range of ports to check is 3000 to 8000.
To search through a wider range of ports, use
\code{\link[httpuv:randomPort]{httpuv::randomPort()}} directly.
}
\details{
This function automatically excludes some ports
which are considered unsafe by web browsers.
}
\examples{

randomPort()
randomPort()
randomPort(cache = FALSE)
}
\seealso{
\code{\link[httpuv:randomPort]{httpuv::randomPort()}}
}
