% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactives.R
\name{reactiveConsole}
\alias{reactiveConsole}
\title{Activate reactivity in the console}
\usage{
reactiveConsole(enabled)
}
\arguments{
\item{enabled}{Turn console reactivity on or off?}
}
\description{
This is an experimental feature that allows you to enable reactivity
at the console, for the purposes of experimentation and learning.
}
\examples{
reactiveConsole(TRUE)
x <- reactiveVal(10)
y <- observe({
  message("The value of x is ", x())
})
x(20)
x(30)
reactiveConsole(FALSE)
}
\keyword{internal}
