% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactives.R
\name{reactiveValuesFromList}
\alias{reactiveValuesFromList}
\title{Convert a list to a reactivevalues object}
\usage{
reactiveValuesFromList(x)
}
\arguments{
\item{x}{A list.}
}
\description{
This function turns a list into a reactivevalues object.
}
\examples{
values <- list(a = 1)
\dontrun{
reactiveValuesFromList(values)
}

}
