% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{key_missing}
\alias{is.key_missing}
\alias{tags}
\alias{p}
\alias{h1}
\alias{h2}
\alias{h3}
\alias{h4}
\alias{h5}
\alias{h6}
\alias{a}
\alias{br}
\alias{div}
\alias{span}
\alias{pre}
\alias{code}
\alias{img}
\alias{strong}
\alias{em}
\alias{hr}
\alias{tag}
\alias{tagList}
\alias{tagAppendAttributes}
\alias{tagHasAttribute}
\alias{tagGetAttribute}
\alias{tagAppendChild}
\alias{tagAppendChildren}
\alias{tagSetChildren}
\alias{HTML}
\alias{includeHTML}
\alias{includeText}
\alias{includeMarkdown}
\alias{includeCSS}
\alias{includeScript}
\alias{singleton}
\alias{is.singleton}
\alias{validateCssUnit}
\alias{htmlTemplate}
\alias{suppressDependencies}
\alias{withTags}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{fastmap}{\code{\link[fastmap:is.key_missing]{is.key_missing()}}, \code{\link[fastmap:key_missing]{key_missing()}}}

  \item{htmltools}{\code{\link[htmltools:a]{a()}}, \code{\link[htmltools:br]{br()}}, \code{\link[htmltools:code]{code()}}, \code{\link[htmltools:div]{div()}}, \code{\link[htmltools:em]{em()}}, \code{\link[htmltools:h1]{h1()}}, \code{\link[htmltools:h2]{h2()}}, \code{\link[htmltools:h3]{h3()}}, \code{\link[htmltools:h4]{h4()}}, \code{\link[htmltools:h5]{h5()}}, \code{\link[htmltools:h6]{h6()}}, \code{\link[htmltools:hr]{hr()}}, \code{\link[htmltools:HTML]{HTML()}}, \code{\link[htmltools:htmlTemplate]{htmlTemplate()}}, \code{\link[htmltools:img]{img()}}, \code{\link[htmltools:includeCSS]{includeCSS()}}, \code{\link[htmltools:includeHTML]{includeHTML()}}, \code{\link[htmltools:includeMarkdown]{includeMarkdown()}}, \code{\link[htmltools:includeScript]{includeScript()}}, \code{\link[htmltools:includeText]{includeText()}}, \code{\link[htmltools:is.singleton]{is.singleton()}}, \code{\link[htmltools:p]{p()}}, \code{\link[htmltools:pre]{pre()}}, \code{\link[htmltools:singleton]{singleton()}}, \code{\link[htmltools:span]{span()}}, \code{\link[htmltools:strong]{strong()}}, \code{\link[htmltools:suppressDependencies]{suppressDependencies()}}, \code{\link[htmltools:tag]{tag()}}, \code{\link[htmltools:tagAppendAttributes]{tagAppendAttributes()}}, \code{\link[htmltools:tagAppendChild]{tagAppendChild()}}, \code{\link[htmltools:tagAppendChildren]{tagAppendChildren()}}, \code{\link[htmltools:tagGetAttribute]{tagGetAttribute()}}, \code{\link[htmltools:tagHasAttribute]{tagHasAttribute()}}, \code{\link[htmltools:tagList]{tagList()}}, \code{\link[htmltools:tags]{tags}}, \code{\link[htmltools:tagSetChildren]{tagSetChildren()}}, \code{\link[htmltools:validateCssUnit]{validateCssUnit()}}, \code{\link[htmltools:withTags]{withTags()}}}
}}

