% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{removeEmptyUsers}
\alias{removeEmptyUsers}
\title{Remove from the users.}
\usage{
removeEmptyUsers(issuesDF, usersDF)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Remove from the users.
}
\examples{
add(1, 1)
add(10, 1)
}

