% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-bootstrap-table.R
\name{renderBootstrapTable}
\alias{renderBootstrapTable}
\title{Table Output (Bootstrap)}
\usage{
renderBootstrapTable(expr, ..., format = "basic", width = "auto",
  env = parent.frame(), quoted = FALSE, func = NULL)
}
\arguments{
\item{expr}{An expression that returns an R object that can be used with
\code{\link[xtable]{xtable}}.}

\item{...}{Arguments to be passed through to \code{\link[xtable]{xtable}} and
\code{\link[xtable]{print.xtable}}.}

\item{format}{An optional string with the Bootstrap table format to apply to the
table (options: striped, bordered, hover, condensed).}

\item{width}{An optional string with the width of the table, as a percentage of
the total width of the page.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}

\item{func}{A function that returns an R object that can be used with
\code{\link[xtable]{xtable}} (deprecated; use \code{expr} instead).}
}
\description{
Creates a reactive table using the Bootstrap design that is suitable for assigning to
an \code{output} slot.
}
\details{
The corresponding HTML output tag should be \code{div} and have the CSS class
name \code{shiny-html-output}.
}

