% Generated by roxygen2 (4.0.0): do not edit by hand
\name{renderError}
\alias{renderError}
\title{Signal a custom type of error when the input values do not satisfy the output
rendering function}
\usage{
renderError(...)
}
\arguments{
\item{...}{A list of arguments, and each argument takes a list of length 2:
the first element of the list is an error message (a character vector), and
the second element is the condition on which the rendering function should
stop (the condition returns \code{TRUE} or \code{FALSE}, and this function
stops when the condition is \code{TRUE})}
}
\description{
For an output rendering function (e.g. \code{\link{renderPlot}()}), we may
need to check certain input values before we can render the output. If the
input values do not satisfy the rendering function, we can siginal an error
to indicate this special situation.
}
\examples{
# in ui.R
fluidPage(
  checkboxGroupInput('foo', 'Check some letters', choices = head(LETTERS)),
  selectizeInput('bar', 'Select a state', choices = state.name),
  plotOutput('plot')
)

# in server.R
function(input, output) {
  output$plot <- renderPlot({
    renderError(
      list(message = 'Check at least one letter!', condition = is.null(input$foo)),
      list(message = 'Please choose a state.', condition = input$bar == '')
    )
    plot(1:10, main = paste(c(input$bar, input$foo), collapse = ', '))
  })
}
}

