% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-input.R
\name{resetBrush}
\alias{resetBrush}
\title{Resets/clears a brush (or brushes) on the client}
\usage{
resetBrush(session = getDefaultReactiveDomain(), brushId)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}. Default is \code{getDefaultReactiveDomain()}.}

\item{brushId}{The id of a brush given to \code{plotOutput} or \code{imageOutput}.}
}
\description{
Sends a message to the client to reset all brushes with the given id.
}
\details{
Multiple brushes from different plot or image outputs may share the
same id, but only one may have an active selection at a time. This
function will reset whichever one is active (if any).

Since brushes are not standard bound inputs, the reset message does not
go onto the input message queue and may be sent immediately.
}
