% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmark-state.R
\name{restoreInput}
\alias{restoreInput}
\title{Restore an input value}
\usage{
restoreInput(id, default)
}
\arguments{
\item{id}{Name of the input value to restore.}

\item{default}{A default value to use, if there's no value to restore.}
}
\description{
This restores an input value from the current restore context. It should be
called early on inside of input functions (like \code{\link[=textInput]{textInput()}}).
}
