% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{result_row}
\alias{result_row}
\title{Creates and returns run result data frame.}
\usage{
result_row(file, pass, result, error)
}
\arguments{
\item{file}{Name of the test runner file, a character vector of length 1.}

\item{pass}{Whether or not the test passed, a logical vector of length 1.}

\item{result}{Value (wrapped in a list) obtained by evaluating \code{file} or \code{NA}
if no value was obtained, such as with \code{shinytest}.}

\item{error}{Error, if any, (and wrapped in a list) that was signaled during
evaluation of \code{file}.}
}
\value{
A 1-row data frame representing a single test run. \code{result} and
\code{error} are "list columns", or columns that may contain list elements.
}
\description{
Creates and returns run result data frame.
}
