% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFeatureFunction.R
\name{RI.length.feature}
\alias{RI.length.feature}
\title{get length of exons and introns associated with RI event}
\usage{
RI.length.feature(event)
}
\arguments{
\item{event}{a event id}
}
\value{
data.frame
}
\description{
get length of exons and introns associated with RI event
}
\keyword{internal}
