% Generated by roxygen2 (4.0.0): do not edit by hand
\name{runRmdContainer}
\alias{runRmdContainer}
\title{Run R Markdown docs with embedded Shiny apps}
\usage{
runRmdContainer(input, text = NULL, ..., knit.options = list())
}
\arguments{
\item{input}{Path to .Rmd file}

\item{text}{A character vector as an alternate way to provide input}

\item{...}{Additional parameters to pass to \code{\link{runApp}}}

\item{knit.options}{A list of options to pass to \code{knit2html}}
}
\description{
Experimental.
}

