% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{saveHdf5File}
\alias{saveHdf5File}
\title{Save R variable to hdf5 file}
\usage{
saveHdf5File(con, dataset, con.type = "F")
}
\arguments{
\item{con}{path to hdf5 file}

\item{dataset}{can be character, numeric, matrix, data.frame, list}

\item{con.type}{connection type}
}
\description{
Save R variable to hdf5 file
}
\keyword{internal}
