% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-state.R
\name{saveStateButton}
\alias{saveStateButton}
\title{Create a button for bookmarking/sharing}
\usage{
saveStateButton(inputId, label = "Save and share...",
  icon = shiny::icon("link", lib = "glyphicon"),
  title = "Save this application's current state and get a URL for sharing.",
  ...)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button or link--usually a text label, but
you could also use any other HTML, like an image.}

\item{icon}{An optional \code{\link{icon}} to appear on the button.}

\item{title}{A tooltip that is shown when the mouse cursor hovers over the
button.}

\item{...}{Named attributes to be applied to the button or link.}
}
\description{
A \code{bookmarkButton} is a \code{\link{actionButton}} with a default label
that consists of a link icon and the text "Share...". It is meant to be used
for bookmarking state.
}
\seealso{
configureBookmarking
}

