% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-state.R
\name{saveStateQueryString}
\alias{encodeStateQueryString}
\alias{saveStateQueryString}
\title{Save or encode state of Shiny session}
\usage{
saveStateQueryString(input, exclude = NULL)

encodeStateQueryString(input, exclude = NULL)
}
\arguments{
\item{input}{The session's input object.}

\item{exclude}{A character vector of input names that should not be
bookmarked.}
}
\description{
Shiny applications can have their state \emph{encoded} in a URL or
\emph{saved}. If the state is encoded, all of the input values are stored in
the URL. If the state is saved, the input values and any uploaded files are
stored on disk.
}

