% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{serverInfo}
\alias{serverInfo}
\title{Collect information about the Shiny Server environment}
\usage{
serverInfo()
}
\value{
A list of the Shiny Server information.
}
\description{
This function returns the information about the current Shiny Server, such as
its version, and whether it is the open source edition or professional
edition. If the app is not served through the Shiny Server, this function
just returns \code{list(shinyServer = FALSE)}.
}
\details{
This function will only return meaningful data when using Shiny Server
version 1.2.2 or later.
}
