\name{serviceApp}
\alias{serviceApp}
\title{Run an application that was created by \code{\link{startApp}}. This
function should normally be called in a \code{while(T)} loop.}
\usage{
  serviceApp(ws_env)
}
\arguments{
  \item{ws_env}{The return value from
  \code{\link{startApp}}.}
}
\description{
  Run an application that was created by
  \code{\link{startApp}}. This function should normally be
  called in a \code{while(T)} loop.
}

