% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmark-state.R
\name{setBookmarkEvent}
\alias{setBookmarkEvent}
\title{Set the event that triggers bookmarking}
\usage{
setBookmarkEvent(expr, env = parent.frame(), quoted = FALSE,
  session = getDefaultReactiveDomain())
}
\description{
Set the event that triggers bookmarking
}

