% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmark-state.R
\name{setBookmarkExclude}
\alias{setBookmarkExclude}
\title{Exclude inputs from bookmarking}
\usage{
setBookmarkExclude(names = character(0),
  session = getDefaultReactiveDomain())
}
\arguments{
\item{names}{A character vector containing names of inputs to exclude from
bookmarking.}
}
\description{
Exclude inputs from bookmarking
}

