% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{setStoryPoints}
\alias{setStoryPoints}
\title{Assigns the same story points to all subtask which does not have.}
\usage{
setStoryPoints(issues)
}
\description{
Assigns the same story points to all subtask which does not have.
}

