% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{shiny.appobj}
\alias{shiny.appobj}
\alias{as.shiny.appobj}
\alias{as.shiny.appobj.shiny.appobj}
\alias{as.shiny.appobj.list}
\alias{as.shiny.appobj.character}
\alias{is.shiny.appobj}
\alias{print.shiny.appobj}
\alias{as.tags.shiny.appobj}
\title{Shiny App object}
\usage{
as.shiny.appobj(x)

\method{as.shiny.appobj}{shiny.appobj}(x)

\method{as.shiny.appobj}{list}(x)

\method{as.shiny.appobj}{character}(x)

is.shiny.appobj(x)

\method{print}{shiny.appobj}(x, ...)

\method{as.tags}{shiny.appobj}(x, ...)
}
\arguments{
\item{x}{Object to convert to a Shiny app.}

\item{...}{Additional parameters to be passed to print.}
}
\description{
Internal methods for the \code{shiny.appobj} S3 class.
}
\keyword{internal}
