% Generated by roxygen2 (4.0.1): do not edit by hand
\name{shiny_demo}
\alias{shiny_demo}
\title{Run a Shiny App from a Package}
\usage{
shiny_demo(topic, package = NULL, lib.loc = NULL,
  verbose = getOption("verbose"), includeAllInstalled = FALSE)
}
\arguments{
\item{topic}{the topic/app which should be run.}

\item{package}{the package which contains the app to run. If \code{NULL} the
first app with the given topic name will be run.}

\item{lib.loc}{a character vector of directory names of R libraries, or NULL.
The default value of NULL corresponds to all libraries currently known. If
the default is used, the loaded packages are searched before the libraries.}

\item{verbose}{a logical. If TRUE, additional diagnostics are printed.}

\item{includeAllInstalled}{a logical. If TRUE and topic not specified, all
Shiny apps from all installed packages will be listed.}
}
\value{
a data frame with two columns, \code{package} and \code{app}, listing
  all the available Shiny apps.
}
\description{
\code{shiny_demo} is a user-friendly interface to running Shiny applications
from R packages. It is designed to work in a similar fashion as the
\code{\link{demo}} function. \code{shiny_demo} gives a list of available
Shiny apps within laoded packages. Setting the \code{includeAllInstalled}
parameter to \code{TRUE} will also search all installed packages but note
it will take longer to execute.
}
\details{
For package developers, simply put Shiny apps in the
\code{inst/shiny} directory of the package. This function will find any
apps located there once installed.
}
\author{
Jason Bryer (jason@bryer.org)
}
\seealso{
demo
}

