% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev-mode.R
\name{shiny_dev_mode}
\alias{shiny_dev_mode}
\title{Enable Shiny developer mode}
\usage{
shiny_dev_mode(
  enable = getOption("shiny.devmode", NULL),
  verbose = getOption("shiny.devmode.verbose", TRUE)
)
}
\arguments{
\item{enable}{Value to set all Shiny developer settings. See details for logical values. If \code{NULL}, no action will be taken.
\code{\link[=options]{options()}} set when \code{enable} is \code{TRUE}:
\itemize{
\item \code{shiny.devmode = TRUE}: Enable Shiny developer mode
\item \code{shiny.launch.browser = TRUE}: Always launch the shiny application window
\item \code{shiny.minified = FALSE}: Use \code{shiny.js} for easier debugging
\item \code{shiny.autoreload = TRUE}: Autoreload when \verb{R/} files change
\item \code{shiny.fullstacktrace = TRUE}: Display full error stack traces
\item \code{sass.cache = FALSE }: Disable default \code{sass} caching
}

If \code{enable} is \code{FALSE}, all \code{\link[=options]{options()}} described above will be set to \code{NULL} to allow for default behavior.}

\item{verbose}{Value to enable printing output. To disable message output, set to \code{FALSE}.}
}
\description{
Sets many options to be used by Shiny application developers.
}
\examples{

# To enable Shiny dev mode for an application,
# place this line of code near the top of your Shiny application
\dontrun{shiny_dev_mode(TRUE)}

# To enable Shiny dev mode for all locally ran applications,
# place this line of code within your `.Rprofile`.
\dontrun{options(shiny.devmode = TRUE)}
# (By only setting the option, `shiny` will not be loaded for every R session)
}
