% Generated by roxygen2 (4.0.0): do not edit by hand
\name{shinyApp}
\alias{shinyApp}
\title{Create a Shiny app object}
\usage{
shinyApp(ui, server, onStart = NULL, options = list())
}
\arguments{
\item{ui}{The UI definition of the app (for example, a call to
\code{fluidPage()} with nested controls)}

\item{server}{A server function}

\item{appDir}{Path to directory that contains a Shiny app (i.e. a server.R
file and either ui.R or www/index.html)}

\item{onStart}{A function that will be called before the app is actually run.
This is only needed for \code{shinyAppObj}, since in the \code{shinyAppDir}
case, a \code{global.R} file can be used for this purpose.}

\item{options}{Named options that should be passed to the `runApp` call. You
can also specify \code{width} and \code{height} parameters which provide a
hint to the embedding environment about the ideal height/width for the app.}
}
\value{
An object that represents the app. Printing the object will run the
  app.
}
\description{
These functions create Shiny app objects from either an explicit UI/server
pair (\code{shinyApp}), or by passing the path of a directory that
contains a Shiny app (\code{shinyAppDir}). You generally shouldn't need to
use these functions to create/run applications; they are intended for
interoperability purposes, such as embedding Shiny apps inside a \pkg{knitr}
document.
}
\examples{
\dontrun{
shinyApp(
  ui = fluidPage(
    numericInput("n", "n", 1),
    plotOutput("plot")
  ),
  server = function(input, output) {
    output$plot <- renderPlot( plot(head(cars, input$n)) )
  },
  options=list(launch.browser = rstudio::viewer)
)

shinyAppDir(system.file("examples/01_hello", package="shiny"))
}
}

