% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{shinyDeprecated}
\alias{shinyDeprecated}
\title{Print message for deprecated functions in Shiny}
\usage{
shinyDeprecated(
  version,
  what,
  with = NULL,
  details = NULL,
  type = c("deprecated", "superseded")
)
}
\arguments{
\item{version}{Shiny version when the function was deprecated}

\item{what}{Function with possible arguments}

\item{with}{Possible function with arguments that should be used instead}

\item{details}{Additional information to be added after a new line to the displayed message}
}
\description{
To disable these messages, use \code{options(shiny.deprecation.messages=FALSE)}.
}
\keyword{internal}
