% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-options.R
\name{getShinyOption}
\alias{getShinyOption}
\alias{setShinyOption}
\alias{shinyOptions}
\title{Get or set Shiny options}
\usage{
getShinyOption(name, default = NULL)

setShinyOption(...)

shinyOptions()
}
\arguments{
\item{name}{Name of an option to get}

\item{default}{Value to be returned if the option is not currently set.}

\item{...}{Options to set, with the form \code{name = value}.}
}
\description{
\code{getShinyOption} retrieves the value of a Shiny option.
\code{setShinyOption} sets the value of a Shiny option. \code{shinyOptions}
returns a list of all currently-set Shiny options.
}
\details{
Shiny stores its options in a stack of \emph{option frames}. There is a
global option frame, which is always available. When an app is run with
\code{\link{runApp}}, another option frame is pushed onto the stack. And
finally, inside an application's server function, another option frame is
pushed on the stack.
}

