% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{shinyPackageSkeleton}
\alias{shinyPackageSkeleton}
\title{Generate a Shiny package template}
\usage{
shinyPackageSkeleton(name, dir, twoFiles = FALSE, rDir = TRUE, examples = TRUE)
}
\description{
Generate a Shiny package template
}
