% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyShortcut.R
\name{shinyShortcut}
\alias{shinyShortcut}
\title{Create Shiny App Shortcut}
\usage{
shinyShortcut(shinyDirectory = getwd(), OS = .Platform$OS.type)
}
\arguments{
\item{shinyDirectory}{The top-level directory containing shiny code and
materials. Also where the final shortcut file will be saved.}

\item{OS}{The operating system that R is being ran on.}
}
\description{
Creates an executable file that, when ran, will open the co-located shiny
app using the user's default internet browser.
}
\details{
Calling \code{shinyShortcut} will write an executable file
that when called will run the shiny app in the default browser.
This executable is stored in a created folder named  .shiny_run'
whilst a shortcut that will silently run the executable
is saved in the home directory as 'shinyShortcut' (this will
be a .vbs script on windows or a .desktop file on unix systems).
}
\examples{
shinyShortcut()
}
