% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{showModal}
\alias{removeModal}
\alias{showModal}
\title{Show a modal dialog}
\usage{
showModal(ui, session = getDefaultReactiveDomain())

removeModal(session = getDefaultReactiveDomain())
}
\arguments{
\item{ui}{UI content to show in the modal.}

\item{session}{The \code{session} object passed to function given to
\code{shinyServer}.}
}
\description{
Show a modal dialog
}
\seealso{
\code{\link{modalDialog}} for examples.
}

