% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snapshotExclude}
\alias{snapshotExclude}
\title{Mark an output to be excluded from test snapshots}
\usage{
snapshotExclude(x)
}
\arguments{
\item{x}{A reactive which will be assigned to an output.}
}
\description{
Mark an output to be excluded from test snapshots
}
