% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snapshotPreprocessInput}
\alias{snapshotPreprocessInput}
\title{Add a function for preprocessing an input before taking a test snapshot}
\usage{
snapshotPreprocessInput(inputId, fun, session = getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{Name of the input value.}

\item{fun}{A function that takes the input value and returns a modified
value. The returned value will be used for the test snapshot.}

\item{session}{A Shiny session object.}
}
\description{
Add a function for preprocessing an input before taking a test snapshot
}
