% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snapshotPreprocessOutput}
\alias{snapshotPreprocessOutput}
\title{Add a function for preprocessing an output before taking a test snapshot}
\usage{
snapshotPreprocessOutput(x, fun)
}
\arguments{
\item{x}{A reactive which will be assigned to an output.}

\item{fun}{A function that takes the output value as an input and returns a
modified value. The returned value will be used for the test snapshot.}
}
\description{
Add a function for preprocessing an output before taking a test snapshot
}
