% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{shinySoftDeprecated}
\alias{shinySoftDeprecated}
\alias{shinySoftDeprecatedMessage}
\title{Print message for deprecated functions in Shiny}
\usage{
shinySoftDeprecated(
  when,
  what,
  with = NULL,
  details = NULL,
  id = NULL,
  env = rlang::caller_env(1)
)

shinySoftDeprecatedMessage(...)
}
\arguments{
\item{when}{The package version when function/argument was deprecated.}

\item{what}{If the deprecated feature is a whole function, the
function name: \code{"foo()"}. If it's an argument that is being
deprecated, the function call should include the argument:
\code{"foo(arg = )"}.

You can optionally supply the namespace: \code{"ns::foo()"}. If not
supplied, it is inferred from the caller environment.}

\item{with}{An optional replacement for the deprecated feature.
This should be a string of the same form as \code{what}.}

\item{details}{The deprecation message is generated from \code{when},
\code{what}, and \code{with}. You can additionally supply a string
\code{details} to be appended to the message.}

\item{id}{The id of the deprecation. A warning is issued only once
for each \code{id}. Defaults to the generated message, but you should
give a unique ID when the message in \code{details} is built
programmatically and depends on inputs, or when you'd like to
deprecate multiple functions but warn only once for all of them.}

\item{env}{The environment in which the deprecated function
was called. A warning is issued if called from the global
environment. If testthat is running, a warning is also called if
the deprecated function was called from the package being tested.

This typically doesn't need to be specified, unless you call
\code{deprecate_soft()} or \code{deprecate_warn()} from an internal helper.
In that case, you need to forward the calling environment.}
}
\description{
To disable these messages, use \code{options(shiny.deprecation.messages=FALSE)}.
}
\section{Functions}{
\itemize{
\item \code{shinySoftDeprecated}: Passes arguments to \code{\link[lifecycle:deprecate_soft]{lifecycle::deprecate_soft()}}

\item \code{shinySoftDeprecatedMessage}: Passes along a message directly to \code{\link[rlang:abort]{rlang::warn()}} with \code{.frequency = "regular"}
}}

\keyword{internal}
