% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otel.R
\name{start_ospan}
\alias{start_ospan}
\title{Start an OpenTelemetry span and return a closure to end it}
\usage{
start_ospan(name, ..., attributes = NULL, domain = getDefaultReactiveDomain())
}
\arguments{
\item{name}{Character string. The name of the span.}

\item{...}{Additional arguments passed to \code{otel::start_span()}.}

\item{domain}{The reactive domain to associate with the span. Defaults to
the current reactive domain from \code{getDefaultReactiveDomain()}.}
}
\value{
A function that, when called, ends the span. The function returns
\code{invisible()} and is safe to call multiple times.
}
\description{
Creates an OpenTelemetry span for discontiguous operations where you need
manual control over when the span ends. Returns a closure function that
can be called to end the span. The returned function is idempotent - it
can be called multiple times but will only end the span once.
}
\examples{
\dontrun{
# Start a span for discontiguous work
end_my_operation_span <- start_ospan("my_operation")
# ... do some work ...
end_my_operation_span()
}

}
\seealso{
\code{\link{with_ospan_async}} for continuous span operations
}
\keyword{internal}
