% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{storyPoints}
\alias{storyPoints}
\title{Assigns the same story points to a subtask which does not have.}
\usage{
storyPoints(key, i)
}
\description{
Assigns the same story points to a subtask which does not have.
}

