% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html-deps.R
\name{suppressDependencies}
\alias{suppressDependencies}
\title{Suppress web dependencies}
\usage{
suppressDependencies(...)
}
\arguments{
\item{...}{Names of the dependencies to suppress. For example,
\code{"jquery"} or \code{"bootstrap"}.}
}
\description{
This suppresses one or more web dependencies. It is meant to be used with
Shiny applications where a dependency (like a JavaScript or CSS file) is
declared in HTML, in an HTML template.
}
\seealso{
\code{\link{htmlTemplate}} for more information about using HTML
  templates with Shiny.

\code{\link[htmltools]{htmlDependency}}
}

