% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html-deps.R
\name{suppressDependency}
\alias{suppressDependency}
\title{Suppress a web dependency}
\usage{
suppressDependency(name)
}
\arguments{
\item{name}{Name of the dependency. For example, \code{"jquery"} or
\code{"bootstrap"}.}
}
\description{
This suppresses a web dependency. It is meant to be used with Shiny
applications where dependency (like a JavaScript or CSS file) is declared in
HTML, in an HTML template.
}
\seealso{
\code{\link{htmlTemplate}} for more information about using HTML
  templates with Shiny.

\code{\link[htmltools]{htmlDependency}}
}

