% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynapseReactClient.R
\name{SynapseReactClient}
\alias{SynapseReactClient}
\title{<Add Title>}
\usage{
SynapseReactClient(
  data = NULL,
  src = c("UserSearchBox"),
  ...,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
<Add Description>
}
