% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{testApp}
\alias{testApp}
\title{Runs the tests associated with this Shiny app}
\usage{
testApp(appDir = ".")
}
\description{
Sources the \code{.R} files in the top-level of \code{tests/} much like \code{R CMD check}.
These files are typically simple runners for tests nested in other
directories under \code{tests/}.
}
\details{
TODO: add in filters for specific tests, or high-level toggles for particular types of tests. Or should that just be a filter on the runners?
TODO: make shinytest suggestable. Right now we'd have to import it.

First checks to see if the \code{.R`` files in the }tests/` directory are
all \href{https://rstudio.github.io/shinytest/}{shinytest}s; if so, just calls out
to \code{\link[shinytest:testApp]{shinytest::testApp()}}.
}
