% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-module.R
\name{testModule}
\alias{testModule}
\title{Test a shiny module}
\usage{
testModule(module, expr, args, ...)
}
\arguments{
\item{module}{The module to test}

\item{expr}{Test code containing expectations. The test expression will run
in the module's environment, meaning that the module's parameters (e.g.
\code{input}, \code{output}, and \code{session}) will be available along with any other
values created inside of the module.}

\item{args}{A list of arguments to pass into the module beyond \code{input},
\code{output}, and \code{session}.}

\item{...}{Additional named arguments to be passed on to the module function.}
}
\description{
Test a shiny module
}
