% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-textsubmit.R
\name{textSubmitInput}
\alias{textSubmitInput}
\title{Create a textarea input control with a submit button}
\usage{
textSubmitInput(
  inputId,
  label,
  value = "",
  ...,
  width = NULL,
  placeholder = NULL,
  priority = c("event", "input"),
  clearOnSubmit = TRUE
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value.}

\item{...}{Ignored, included to require named arguments and for future
feature expansion.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[=validateCssUnit]{validateCssUnit()}}.}

\item{placeholder}{A character string giving the user a hint as to what can be
entered into the control.}

\item{priority}{The priority of the input. If \code{"event"}, the server input value
will always invalidate}

\item{clearOnSubmit}{Whether the input should be cleared after submitting.}
}
\value{
A textarea input control that can be added to a UI definition.
}
\description{
Create a textarea input control for entry of unstructured text values.
The input includes a submit button that allows users to explicitly submit their input.
This is useful for scenarios where you want to capture the final input rather than reacting to every keystroke.
}
\section{Server value}{

A character string of the text input. The default value is \code{""}
unless \code{value} is provided.
}

\examples{
## Only run examples in interactive R sessions
if (interactive()) {

ui <- fluidPage(
  textSubmitInput("text", "textSubmitInput()", placeholder = "Enter some input..."),
  verbatimTextOutput("value")
)
server <- function(input, output) {
  output$value <- renderText({ input$text })
}
shinyApp(ui, server)

}

}
\seealso{
\code{\link[=updateTextAreaInput]{updateTextAreaInput()}}

Other input elements: 
\code{\link{actionButton}()},
\code{\link{checkboxGroupInput}()},
\code{\link{checkboxInput}()},
\code{\link{dateInput}()},
\code{\link{dateRangeInput}()},
\code{\link{fileInput}()},
\code{\link{numericInput}()},
\code{\link{passwordInput}()},
\code{\link{radioButtons}()},
\code{\link{selectInput}()},
\code{\link{sliderInput}()},
\code{\link{submitButton}()},
\code{\link{textAreaInput}()},
\code{\link{textInput}()},
\code{\link{varSelectInput}()}
}
\concept{input elements}
