% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/image-interact.R
\name{underBrush}
\alias{underBrush}
\title{Find points that are under a brush}
\usage{
underBrush(brush, df, xvar, yvar)
}
\arguments{
\item{brush}{The data from a brush, such as \code{input$plot_brush}.}

\item{df}{A data frame from which to select rows.}

\item{xvar}{A string with the name of the variable on the x axis. This must
also be the name of a column in \code{df}.}

\item{yvar}{A string with the name of the variable on the y axis. This must
  also be the name of a column in \code{df}.}
}
\description{
This function returns rows from a data frame which are under a brush used
with \code{\link{plotOutput}}.
}
\details{
If a specified x or y column is a factor, then it will be coerced to an
integer vector. If it is a character vector, then it will be coerced to a
factor and then integer vector. This means that the brush will be considered
to cover a given character/factor value when it covers the center value.
}
\seealso{
\code{\link{plotOutput}} for example usage.
}

