% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-input.R
\name{updateActionButton}
\alias{updateActionButton}
\title{Change the label or icon of an action button on the client}
\usage{
updateActionButton(session, inputId, label = NULL, icon = NULL)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{icon}{The icon to set for the input object. To
remove an existing icon, use \code{icon=icon("")}.}
}
\description{
Change the label or icon of an action button on the client
}
\details{
The input updater functions send a message to the client, telling it to
change the settings of an input object. The messages are collected and sent
after all the observers (including outputs) have finished running.

The syntax of these functions is similar to the functions that created the
inputs in the first place. For example, \code{\link{numericInput}()} and
\code{updateNumericInput()} take a similar set of arguments.

Any arguments with NULL values will be ignored; they will not result in any
changes to the input object on the client.

For \code{\link{radioButtons}()}, \code{\link{checkboxGroupInput}()} and
\code{\link{selectInput}()}, the set of choices can be cleared by using
\code{choices=character(0)}
}
\examples{
\dontrun{
shinyServer(function(input, output, session) {

  observe({
    updateActionButton(session, "goButton",
      label = "New label",
      icon = icon("calendar"))
  })
})
}
}
\seealso{
\code{\link{actionButton}}
}

