% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-input.R
\name{updateBrushCoords}
\alias{updateBrushCoords}
\title{Change the region selected by a brush on the client}
\usage{
updateBrushCoords(
  session = getDefaultReactiveDomain(),
  brushId,
  coords,
  panel = 0,
  outputId = NULL
)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}. Default is \code{getDefaultReactiveDomain()}.}

\item{brushId}{The id of a brush given to \code{plotOutput} or \code{imageOutput}.}

\item{coords}{A list containing \code{xmin}, \code{xmax}, \code{ymin}, \code{ymax} numerics.}

\item{panel}{The numeric index of the panel to brush. Indices start from 0,
which is the default, and the only valid value for single-panel plots.}

\item{outputId}{The id of the \code{plotOutput} or \code{imageOutput} that contains
the brush to update. Can be NULL (default) if the brushId is unique.}
}
\description{
Sends a message to the client to set the region selected by a particular
plot/image brush to the specified coordinates.
}
\details{
Sets the brush with the given \code{brushId}, if it exists on an output with
the given \code{outputId} (or, if none is specified, any \code{imageOutput} or
\code{plotOutput} in the app). The \code{coords} should be a
list with \code{xmin}, \code{xmax}, \code{ymin}, \code{ymax} single-element numerics that
specify the desired brush position (in the scale of the plot). \code{panel} should
be a single-element integer that defines the panel that should be brushed
for multi-faceted plots.

If the direction of the brush was set to "x" in \code{brushOpts()}, then \code{ymin}
and \code{ymax} will be ignored, but must still be supplied, and vice versa.

\code{outputId} need not be provided if the \code{brushId} is unique. However, if
multiple plot or image outputs have brushes with the same id, this must
be provided to specify which one should be brushed (since no more than one
may have an active brush at one time).

Since brushes are not standard bound inputs, the update message does not
go onto the input message queue and may be sent immediately.
}
\seealso{
\code{\link[=brushOpts]{brushOpts()}}
}
