% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-input.R
\name{updateFileInput}
\alias{updateFileInput}
\title{Change the value of a file input}
\usage{
updateFileInput(session, inputId, label = NULL, value = NULL)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{value}{The file (or files) to set as the value of the file input.}
}
\description{
Change the value of a file input
}
\details{
The input updater functions send a message to the client, telling it to
change the settings of an input object. The messages are collected and sent
after all the observers (including outputs) have finished running.

The syntax of these functions is similar to the functions that created the
inputs in the first place. For example, \code{\link{numericInput}()} and
\code{updateNumericInput()} take a similar set of arguments.

Any arguments with NULL values will be ignored; they will not result in any
changes to the input object on the client.

For \code{\link{radioButtons}()}, \code{\link{checkboxGroupInput}()} and
\code{\link{selectInput}()}, the set of choices can be cleared by using
\code{choices=character(0)}. Similarly, for these inputs, the selected item
can be cleared by using \code{selected=character(0)}.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
  shinyApp(
    ui = fluidPage(
      fileInput("file", "File", multiple = TRUE),
      actionButton("btn_1", "Select the shiny DESCRIPTION"),
      actionButton("btn_2", "Select the shiny DESCRIPTION and NAMESPACE"),
      tableOutput("table")
    ),
    server = function(input, output, session) {
      observeEvent(input$btn_1, {
        updateFileInput(session, "file",
                        value = system.file("DESCRIPTION", package = "shiny"))
      })

      observeEvent(input$btn_2, {
        updateFileInput(session, "file",
                        value = c(system.file("DESCRIPTION", package = "shiny"),
                                  system.file("NAMESPACE", package = "shiny")))
      })

      output$table <- renderTable({
        input$file
      })
    }
  )
}
}
\seealso{
\code{\link{fileInput}}
}
