% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-state.R
\name{updateLocationBar}
\alias{updateLocationBar}
\title{Update URL in browser's location bar}
\usage{
updateLocationBar(queryString, session = getDefaultReactiveDomain())
}
\arguments{
\item{queryString}{The new query string to show in the location bar.}

\item{session}{A Shiny session object.}
}
\description{
This function updates the URL in the client browser's location bar. It
typically is called from an observer.
}

