% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-state.R
\name{updateQueryString}
\alias{updateQueryString}
\title{Update URL query string in browser's location bar}
\usage{
updateQueryString(queryString, session = getDefaultReactiveDomain())
}
\arguments{
\item{queryString}{The new query string to show in the location bar.}

\item{session}{A Shiny session object.}
}
\description{
Update URL query string in browser's location bar
}

