% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmark-state.R
\name{urlModal}
\alias{urlModal}
\title{Generate a modal dialog that displays a URL}
\usage{
urlModal(url, title = "Bookmarked application link", subtitle = NULL)
}
\arguments{
\item{url}{A URL to display in the dialog box.}

\item{title}{A title for the dialog box.}

\item{subtitle}{Text to display underneath URL.}
}
\description{
The modal dialog generated by \code{urlModal} will display the URL in a
textarea input, and the URL text will be selected so that it can be easily
copied. The result from \code{urlModal} should be passed to the
\code{\link[=showModal]{showModal()}} function to display it in the browser.
}
