\name{wellPanel}
\alias{wellPanel}
\title{Create a well panel}
\usage{
  wellPanel(...)
}
\arguments{
  \item{...}{UI elements to include inside the panel.}
}
\value{
  The newly created panel.
}
\description{
  Creates a panel with a slightly inset border and grey
  background. Equivalent to Twitter Bootstrap's
  \code{\href{http://twitter.github.com/bootstrap/components.html#misc}{well}}
  CSS class.
}

