% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otel-with.R
\name{withOtel}
\alias{withOtel}
\title{Set OpenTelemetry options for Shiny reactives}
\usage{
withOtel(expr, ..., bind = "all")
}
\arguments{
\item{expr}{The expression to run with OpenTelemetry spans enabled.}

\item{...}{Future parameter expansion.}

\item{bind}{If \code{"all"} (default), then all reactive objects will be bound to Open Telemetry spans.
If \code{"none"}, then only the reactive objects created within the expression
will be bound to Open Telemetry spans.}
}
\description{
Set OpenTelemetry options for Shiny reactives
}
\examples{
# TODO: Make examples!!
}
