% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-otel.R
\name{withOtelShiny}
\alias{withOtelShiny}
\title{Set OpenTelemetry options for Shiny reactives}
\usage{
withOtelShiny(expr, ..., bindAll = TRUE)
}
\arguments{
\item{expr}{The expression to run with OpenTelemetry spans enabled.}

\item{...}{Future parameter expansion.}

\item{bindAll}{If \code{TRUE}, then all reactive objects will be bound to Open Telemetry spans.
If \code{FALSE}, then only the reactive objects created within the expression
will be bound to Open Telemetry spans.
Defaults to \code{TRUE}.}
}
\description{
Set OpenTelemetry options for Shiny reactives
}
\examples{
# TODO: Update examples!!
}
