% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{worklogOfAssignee}
\alias{worklogOfAssignee}
\title{Returns the worklog of a user.}
\usage{
worklogOfAssignee(assigneeKey, issuesMatrix, worklogsMatrix, start, end)
}
\description{
Returns the worklog of a user.
}

