\name{actionButton}
\alias{actionButton}
\title{Action button}
\usage{
  actionButton(inputId, label)
}
\arguments{
  \item{inputId}{Specifies the input slot that will be used
  to access the value.}

  \item{label}{The contents of the button--usually a text
  label, but you could also use any other HTML, like an
  image.}
}
\description{
  Creates an action button whose value is initially zero,
  and increments by one each time it is pressed.
}

