% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrapLib}
\alias{bootstrapLib}
\title{Bootstrap libraries}
\usage{
bootstrapLib(theme = NULL)
}
\arguments{
\item{theme}{deprecated (it never worked).}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("superseded")}

See \code{\link[bootstraplib:bootstrap]{bootstraplib::bootstrap()}} for a modern alternative for accessing
Bootstrap's HTML dependencies.
}
\keyword{internal}
