% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrapLib}
\alias{bootstrapLib}
\title{Bootstrap libraries}
\usage{
bootstrapLib(theme = NULL)
}
\arguments{
\item{theme}{deprecated (it never worked).}
}
\description{
This function is deprecated in favor of \code{\link[bootstraplib:bootstrap]{bootstraplib::bootstrap()}}, which
provides a more fully featured approach to accessing and theming Bootstrap
CSS. Note that most users won't need to call this function directly since
\code{\link[=bootstrapPage]{bootstrapPage()}} and it's special cases ( e.g., \code{\link[=fluidPage]{fluidPage()}},
\code{\link[=basicPage]{basicPage()}}, \code{\link[=navbarPage]{navbarPage()}}, etc) implicitly depend on Bootstrap.
}
\keyword{internal}
