% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrapLib}
\alias{bootstrapLib}
\title{Bootstrap libraries}
\usage{
bootstrapLib(theme = NULL)
}
\arguments{
\item{theme}{One of the following:
\itemize{
\item \code{NULL} (the default), which implies a "stock" build of Bootstrap 3.
\item A \code{\link[bslib:bs_theme]{bslib::bs_theme()}} object. This can be used to replace a stock
build of Bootstrap 3 with a customized version of Bootstrap 3 or higher.
\item A character string pointing to an alternative Bootstrap stylesheet
(normally a css file within the www directory, e.g. \code{www/bootstrap.css}).
}}
}
\description{
This function defines a set of web dependencies necessary for using Bootstrap
components in a web page.
}
\details{
It isn't necessary to call this function if you use \code{\link[=bootstrapPage]{bootstrapPage()}} or
others which use \code{bootstrapPage}, such \code{\link[=fluidPage]{fluidPage()}}, \code{\link[=navbarPage]{navbarPage()}},
\code{\link[=fillPage]{fillPage()}}, etc, because they already include the Bootstrap web dependencies.
}
