% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-deprecated.R
\name{bootstrapPage}
\alias{bootstrapPage}
\alias{basicPage}
\title{Create a Bootstrap page}
\usage{
bootstrapPage(..., title = NULL, responsive = NULL, theme = NULL)

basicPage(...)
}
\arguments{
\item{...}{The contents of the document body.}

\item{title}{The browser window title (defaults to the host URL of the page)}

\item{responsive}{This option is deprecated; it is no longer optional with
Bootstrap 3.}

\item{theme}{Alternative Bootstrap stylesheet (normally a css file within the
www directory, e.g. \code{www/bootstrap.css})}
}
\value{
A UI defintion that can be passed to the \link{shinyUI} function.
}
\description{
\strong{DEPRECATED}: use \code{\link[=fluidPage]{fluidPage()}} instead.
}
\seealso{
\code{\link[=fluidPage]{fluidPage()}}, \code{\link[=fixedPage]{fixedPage()}}
}
\keyword{internal}
