% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bootstrapPage}
\alias{basicPage}
\alias{bootstrapPage}
\title{Create a Bootstrap page}
\usage{
bootstrapPage(..., title = NULL, responsive = TRUE, theme = NULL)

basicPage(...)
}
\arguments{
\item{...}{The contents of the document body.}

\item{title}{The browser window title (defaults to the host URL of the page)}

\item{responsive}{\code{TRUE} to use responsive layout (automatically adapt
and resize page elements based on the size of the viewing device)}

\item{theme}{Alternative Bootstrap stylesheet (normally a css file within the
  www directory, e.g. \code{www/bootstrap.css})}
}
\value{
A UI defintion that can be passed to the \link{shinyUI} function.
}
\description{
Create a Shiny UI page that loads the CSS and JavaScript for
\href{http://getbootstrap.com/2.3.2/}{Bootstrap}, and has no content in the
page body (other than what you provide).
}
\details{
This function is primarily intended for users who are proficient in
HTML/CSS, and know how to lay out pages in Bootstrap. Most applications
should use \code{\link{fluidPage}} along with layout functions like
\code{\link{fluidRow}} and \code{\link{sidebarLayout}}.
}
\note{
The \code{basicPage} function is deprecated, you should use the
  \code{\link{fluidPage}} function instead.
}
\seealso{
\code{\link{fluidPage}}, \code{\link{fixedPage}}
}

