% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busy-indicators.R
\name{busyIndicatorOptions}
\alias{busyIndicatorOptions}
\title{Customize busy indicator options.}
\usage{
busyIndicatorOptions(
  ...,
  spinner_color = NULL,
  spinner_size = NULL,
  spinner_delay = NULL,
  pulse_background = NULL,
  pulse_height = NULL,
  pulse_speed = NULL
)
}
\arguments{
\item{...}{Currently ignored (for future expansion).}

\item{spinner_color}{The color of the spinner. This can be any valid CSS
color. Defaults to the app's "primary" color if Bootstrap is on the page.}

\item{spinner_size}{The size of the spinner. This can be any valid CSS size.}

\item{spinner_delay}{The amount of time to wait before showing the spinner.
This can be any valid CSS time and can be useful for not showing the spinner
if the computation finishes quickly.}

\item{pulse_background}{A CSS background definition for the pulse. The
default uses a
\href{https://developer.mozilla.org/en-US/docs/Web/CSS/gradient/linear-gradient}{linear-gradient}
of the theme's indigo, purple, and pink colors.}

\item{pulse_height}{The height of the pulsing banner. This can be any valid
CSS size.}

\item{pulse_speed}{The speed of the pulsing banner. This can be any valid CSS
time.}
}
\description{
To customize the appearance of the busy indicators, include the result of
this function in the app's UI.
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

library(bslib)

ui <- page_fillable(
  useBusyIndicators(),
  busyIndicatorOptions(spinner_color = "orange"),
  card(
    card_header(
      "A plot",
      input_task_button("simulate", "Simulate"),
      class = "d-flex justify-content-between align-items-center"
    ),
    plotOutput("p"),
  )
)

server <- function(input, output) {
  output$p <- renderPlot({
    input$simulate
    Sys.sleep(4)
    plot(x = rnorm(100), y = rnorm(100))
  })
}

shinyApp(ui, server)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=useBusyIndicators]{useBusyIndicators()}} for enabling/disabling busy indicators.
}
