% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busy-indicators.R
\name{busyIndicatorOptions}
\alias{busyIndicatorOptions}
\title{Customize busy indicator options.}
\usage{
busyIndicatorOptions(
  ...,
  spinner_type = NULL,
  spinner_color = NULL,
  spinner_size = NULL,
  spinner_delay = NULL,
  spinner_selector = ":root",
  pulse_background = NULL,
  pulse_height = NULL,
  pulse_speed = NULL
)
}
\arguments{
\item{...}{Currently ignored (for future expansion).}

\item{spinner_type}{The type of spinner to use for the busy indicator. This
can be any of the spinner types from the \code{shiny} package.

Available spinner types include:
"12-dots-scale-rotate", "180-ring-with-bg", "180-ring", "270-ring-with-bg", "270-ring", "3-dots-bounce", "3-dots-fade", "3-dots-move", "3-dots-rotate", "3-dots-scale-middle", "3-dots-scale", "6-dots-rotate", "6-dots-scale-middle", "6-dots-scale", "8-dots-rotate", "90-ring-with-bg", "90-ring", "bars-fade", "bars-rotate-fade", "bars-scale-fade", "bars-scale-middle", "bars-scale", "blocks-scale", "blocks-shuffle-2", "blocks-shuffle-3", "blocks-wave", "bouncing-ball", "clock", "dot-revolve", "eclipse-half", "eclipse", "gooey-balls-1", "gooey-balls-2", "pulse-2", "pulse-3", "pulse-multiple", "pulse-ring", "pulse-rings-2", "pulse-rings-3", "pulse-rings-multiple", "pulse", "ring-resize", "tadpole", "wifi-fade", "wifi", "wind-toy".

Alternatively, you can provide a path to a custom spinner SVG file. There
are many free spinner SVGs available online, such as those from
the \href{https://github.com/n3r4zzurr0/svg-spinners}{svg-spinners} repository.}

\item{spinner_color}{The color of the spinner. This can be any valid CSS
color. Defaults to the app's "primary" color if Bootstrap is on the page.}

\item{spinner_size}{The size of the spinner. This can be any valid CSS size.}

\item{spinner_delay}{The amount of time to wait before showing the spinner.
This can be any valid CSS time and can be useful for not showing the spinner
if the computation finishes quickly.}

\item{spinner_selector}{A CSS selector for scoping the spinner customization.
This can be useful if you want to have different spinners for different
parts of the app. Defaults to the root document element. Use \code{NULL} to
return inline styles for spinner customizations (note that you cannot
customize both the pulse and the spinner without providing a selector).}

\item{pulse_background}{A CSS background definition for the pulse. The
default uses a
\href{https://developer.mozilla.org/en-US/docs/Web/CSS/gradient/linear-gradient}{linear-gradient}
of the theme's indigo, purple, and pink colors.}

\item{pulse_height}{The height of the pulsing banner. This can be any valid
CSS size.}

\item{pulse_speed}{The speed of the pulsing banner. This can be any valid CSS
time.}
}
\description{
To customize the appearance of the busy indicators, include the result of
this function in the app's UI.
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

library(bslib)

ui <- page_fillable(
  useBusyIndicators(),
  busyIndicatorOptions(spinner_color = "orange"),
  card(
    card_header(
      "A plot",
      input_task_button("simulate", "Simulate"),
      class = "d-flex justify-content-between align-items-center"
    ),
    plotOutput("p"),
  )
)

server <- function(input, output) {
  output$p <- renderPlot({
    input$simulate
    Sys.sleep(4)
    plot(x = rnorm(100), y = rnorm(100))
  })
}

shinyApp(ui, server)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=useBusyIndicators]{useBusyIndicators()}} for enabling/disabling busy indicators.
}
