% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkboxInput}
\alias{checkboxInput}
\title{Checkbox Input Control}
\usage{
checkboxInput(inputId, label, value = FALSE)
}
\arguments{
\item{inputId}{Input variable to assign the control's value to.}

\item{label}{Display label for the control.}

\item{value}{Initial value (\code{TRUE} or \code{FALSE}).}
}
\value{
A checkbox control that can be added to a UI definition.
}
\description{
Create a checkbox that can be used to specify logical values.
}
\examples{
checkboxInput("outliers", "Show outliers", FALSE)
}
\seealso{
\code{\link{checkboxGroupInput}}, \code{\link{updateCheckboxInput}}

Other input.elements: \code{\link{actionButton}},
  \code{\link{actionLink}}; \code{\link{animationOptions}},
  \code{\link{sliderInput}};
  \code{\link{checkboxGroupInput}};
  \code{\link{dateInput}}; \code{\link{dateRangeInput}};
  \code{\link{fileInput}}; \code{\link{numericInput}};
  \code{\link{radioButtons}}; \code{\link{selectInput}},
  \code{\link{selectizeInput}}; \code{\link{submitButton}};
  \code{\link{textInput}}
}

