% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-interact-opts.R
\name{clickOpts}
\alias{clickOpts}
\title{Create an object representing click options}
\usage{
clickOpts(id, clip = TRUE)
}
\arguments{
\item{id}{Input value name. For example, if the value is \code{"plot_click"},
then the click coordinates will be available as \code{input$plot_click}.}

\item{clip}{Should the click area be clipped to the plotting area? If FALSE,
then the server will receive click events even when the mouse is outside
the plotting area, as long as it is still inside the image.}
}
\description{
This generates an object representing click options, to be passed as the
\code{click} argument of \code{\link[=imageOutput]{imageOutput()}} or
\code{\link[=plotOutput]{plotOutput()}}.
}
