% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-state.R
\name{configureBookmarking}
\alias{configureBookmarking}
\title{Configure bookmarking for the current session}
\usage{
configureBookmarking(eventExpr, enable = TRUE, type = c("save", "encode"),
  exclude = NULL, onBookmarked = NULL,
  session = getDefaultReactiveDomain())
}
\arguments{
\item{eventExpr}{An expression to listen for, similar to
\code{\link{observeEvent}}.}

\item{enable}{If \code{TRUE} (the default), enable bookmarking for this app.}

\item{type}{Either \code{"save"}, which saves to disk, or \code{"encode"},
which encodes all of the relevant values in a URL.}

\item{exclude}{Input values to exclude from bookmarking.}

\item{onBookmarked}{A callback function to invoke after the bookmarking has
been done.}

\item{session}{A Shiny session object.}
}
\description{
There are two types of bookmarking: saving state, and encoding state.
}

