% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmark-state.R
\name{configureBookmarking}
\alias{configureBookmarking}
\title{Configure bookmarking for a Shiny application}
\usage{
configureBookmarking(store = c("url", "server", "disable"), exclude = NULL)
}
\arguments{
\item{store}{Either \code{"url"}, which encodes all of the relevant values in
a URL, \code{"server"}, which saves to disk on the server, or
\code{"disable"}, which disables any previously-enabled bookmarking.}

\item{exclude}{A character vector of names of input values to exclude from
bookmarking.}
}
\description{
Configure bookmarking for a Shiny application
}

