% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinywrappers.R
\name{deprecatedReactives}
\alias{deprecatedReactives}
\alias{reactivePlot}
\alias{reactiveTable}
\alias{reactivePrint}
\alias{reactiveUI}
\alias{reactiveText}
\title{Deprecated reactive functions}
\usage{
reactivePlot(func, width = "auto", height = "auto", ...)

reactiveTable(func, ...)

reactivePrint(func)

reactiveUI(func)

reactiveText(func)
}
\arguments{
\item{func}{A function.}

\item{width}{Width.}

\item{height}{Height.}

\item{...}{Other arguments to pass on.}
}
\description{
\code{reactivePlot} has been replaced by \code{\link[=renderPlot]{renderPlot()}}.

\code{reactiveTable} has been replaced by \code{\link[=renderTable]{renderTable()}}.

\code{reactivePrint} has been replaced by \code{\link[=renderPrint]{renderPrint()}}.

\code{reactiveUI} has been replaced by \code{\link[=renderUI]{renderUI()}}.

\code{reactiveText} has been replaced by \code{\link[=renderText]{renderText()}}.
}
\keyword{internal}
