% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resume.R
\name{enableResume}
\alias{enableResume}
\alias{disableResume}
\title{Resume sessions after an interruption}
\usage{
enableResume(
  reload = c("ask", "resume", "fresh"),
  appVersion = NULL,
  maxSize = 5e+07,
  maxNodeSize = 5e+06,
  debounce = 1000
)

disableResume(x)
}
\description{
Resume sessions after an interruption
}
