% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-lang.R
\name{exprToFunction}
\alias{exprToFunction}
\title{Convert an expression to a function}
\usage{
exprToFunction(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{A quoted or unquoted expression, or a quosure.}

\item{env}{The desired environment for the function. Defaults to the
calling environment two steps back.}

\item{quoted}{Is the expression quoted?}
}
\description{
Similar to \code{\link[=installExprFunction]{installExprFunction()}} but doesn't register debug hooks.
}
\seealso{
\code{\link[=installExprFunction]{installExprFunction()}} for the modern approach to converting an expression to a function
}
