% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{getCurrentTheme}
\alias{getCurrentTheme}
\title{Obtain Shiny's Bootstrap Sass theme}
\usage{
getCurrentTheme()
}
\value{
If called at render-time (i.e., inside a \code{\link[htmltools:tagFunction]{htmltools::tagFunction()}}),
and \code{\link[=bootstrapLib]{bootstrapLib()}}'s \code{theme} has been set to a \code{\link[bslib:bs_theme]{bslib::bs_theme()}}
object, then this returns the \code{theme}. Otherwise, this returns \code{NULL}.
}
\description{
Intended for use by Shiny developers to create Shiny bindings with intelligent
styling based on the \code{\link[=bootstrapLib]{bootstrapLib()}}'s \code{theme} value.
}
\seealso{
\code{\link[=getCurrentOutputInfo]{getCurrentOutputInfo()}}, \code{\link[=bootstrapLib]{bootstrapLib()}}, \code{\link[htmltools:tagFunction]{htmltools::tagFunction()}}
}
\keyword{internal}
