% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{helpText}
\alias{helpText}
\title{Create a help text element}
\usage{
helpText(...)
}
\arguments{
\item{...}{One or more help text strings (or other inline HTML elements)}
}
\value{
A help text element that can be added to a UI definition.
}
\description{
Create help text which can be added to an input form to provide additional
explanation or context.
}
\examples{
helpText("Note: while the data view will show only",
         "the specified number of observations, the",
         "summary will be based on the full dataset.")
}
