\name{horizontalLayout}
\alias{horizontalLayout}
\title{Layout UI elements horizontally}
\usage{
  horizontalLayout(...)
}
\arguments{
  \item{...}{Elements to include within the container}
}
\description{
  Create a container that includes several elements laid
  out side-by-side.
}
\details{
  To force elements to the left or right of the container
  you can use the \code{\link{pullLeft}} and
  \code{\link{pullRight}} functions.
}

