% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{htmlTemplate}
\alias{htmlTemplate}
\title{Process an HTML template}
\usage{
htmlTemplate(filename, ...)
}
\arguments{
\item{filename}{Path to an HTML template file.}

\item{...}{Variable values to use when processing the template.}
}
\description{
Process an HTML template
}

