% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{icon}
\alias{icon}
\title{Create an icon}
\usage{
icon(name, class = NULL, lib = "font-awesome", ...)
}
\arguments{
\item{name}{The name of the icon. Icons are drawn from the \href{https://fontawesome.com/}{Font Awesome Free} (currently icons from the v5.13.0 set are
supported with the v4 naming convention) and
\href{https://getbootstrap.com/docs/3.3/components/#glyphicons}{Glyphicons}
libraries. Note that the \code{"fa-"} and \code{"glyphicon-"} prefixes should not be
used in icon names (i.e. the \code{"fa-calendar"} icon should be referred to as
\code{"calendar"}).}

\item{class}{Additional classes to customize the style of an icon (see the
\href{https://fontawesome.com/how-to-use}{usage examples} for details on
supported styles).}

\item{lib}{The icon library to use. If drawing from the Font Awesome library
of icons, this should be \code{"font-awesome"}. The other option is or
\code{"glyphicon"} to use icons from the Glyphicons library.}

\item{...}{Arguments passed to the \verb{<i>} tag of \link[htmltools:builder]{htmltools::tags}.}
}
\value{
An icon element.
}
\description{
Create an icon for use within a page. Icons can appear on their own, inside
of a button, or as an icon for a \code{\link[=tabPanel]{tabPanel()}} within a \code{\link[=navbarPage]{navbarPage()}}.
}
\examples{
# add an icon to a submit button
submitButton("Update View", icon = icon("redo"))

navbarPage("App Title",
  tabPanel("Plot", icon = icon("bar-chart-o")),
  tabPanel("Summary", icon = icon("list-alt")),
  tabPanel("Table", icon = icon("table"))
)
}
\seealso{
For lists of available icons, see
\url{https://fontawesome.com/icons} and
\url{https://getbootstrap.com/docs/3.3/components/#glyphicons}.
}
