% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{icon}
\alias{icon}
\title{Create an icon}
\usage{
icon(name, class = NULL, lib = c("font-awesome", "fa4", "fa5",
  "glyphicon"))
}
\arguments{
\item{name}{Name of icon. Icons are drawn from the
\href{https://fontawesome.com/}{Font Awesome Free} (currently icons from
version \Sexpr[echo=TRUE,stage=build]{shiny:::fa_version} are supported,
along with backward compatibility for v4 icons) and
\href{http://getbootstrap.com/components/#glyphicons}{Glyphicons}
libraries. Note that the "fa-" and "glyphicon-" prefixes should not be used
in icon names (i.e. the "fa-calendar" icon should be referred to as
"calendar"). May also be \code{NULL}.}

\item{class}{Additional classes to customize the style of the icon (see the
\href{http://fontawesome.io/examples/}{usage examples} for details on
supported styles).}

\item{lib}{Icon library to use. Can be "font-awesome" (the default), "fa5",
"fa4", or "glyphicon". If "font-awesome", that means to use the icon with
the given name from the current version of Font-Awesome
(\Sexpr[echo=TRUE,stage=build]{shiny:::fa_version}), and if it is not found
there, then use the icon from Font-Awesome 4. If "fa5", that means to use
Font-Awesome version 5 only; if "fa4", that means to use Font-Awesome
version 4 only.}
}
\value{
An icon element
}
\description{
Create an icon for use within a page. Icons can appear on their own, inside
of a button, or as an icon for a \code{\link{tabPanel}} within a
\code{\link{navbarPage}}.
}
\examples{
icon("calendar")               # Standard icon, from Font-Awesome 5
icon("calendar", "fa-3x")      # 3x normal size
icon("calendar", lib = "fa4")  # From Font-Awesome 4
icon("battery")                # This will use the "battery" icon from FA v4,
                               # because "battery" is not present in v5.
icon("cog", lib = "glyphicon") # From glyphicon library

# add an icon to a submit button
submitButton("Update View", icon = icon("refresh"))

navbarPage("App Title",
  tabPanel("Plot", icon = icon("bar-chart-o")),
  tabPanel("Summary", icon = icon("list-alt")),
  tabPanel("Table", icon = icon("table"))
)
}
\seealso{
For lists of available icons, see
  \href{http://fontawesome.io/icons/}{http://fontawesome.io/icons/} and
  \href{http://getbootstrap.com/components/#glyphicons}{http://getbootstrap.com/components/#glyphicons}.
}
