% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-layout.R
\name{inputPanel}
\alias{inputPanel}
\title{Input panel}
\usage{
inputPanel(...)
}
\arguments{
\item{...}{Input controls or other HTML elements.}
}
\description{
A \code{\link[=flowLayout]{flowLayout()}} with a grey border and light grey background,
suitable for wrapping inputs.
}
