% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{installExprFunction}
\alias{installExprFunction}
\title{Install an expression as a function}
\usage{
installExprFunction(
  expr,
  name,
  eval.env = parent.frame(2),
  quoted = FALSE,
  assign.env = parent.frame(1),
  label = deparse(sys.call(-1)[[1]]),
  wrappedWithLabel = TRUE,
  ..stacktraceon = FALSE
)
}
\arguments{
\item{expr}{A user-supplied reactive expression (or quosure).}

\item{name}{The name the function should be given}

\item{eval.env}{The desired environment for the function. Defaults to the
calling environment two steps back.}

\item{quoted}{Is \code{expr} quoted?}

\item{assign.env}{The environment in which the function should be assigned.}

\item{label}{A label for the object to be shown in the debugger. Defaults to
the name of the calling function.}

\item{wrappedWithLabel, ..stacktraceon}{Advanced use only. For stack
manipulation purposes; see \code{\link[=stacktrace]{stacktrace()}}.}
}
\description{
Superceded by \code{\link[=exprToFunction]{exprToFunction()}}: use that function instead for a more
intuitive way to pass reactive expressions to \code{\link[=createRenderFunction]{createRenderFunction()}}.
}
\keyword{internal}
