\name{invalidateLater}
\alias{invalidateLater}
\title{Scheduled Invalidation}
\usage{
  invalidateLater(millis)
}
\arguments{
  \item{millis}{Approximate milliseconds to wait before
  invalidating the current reactive context.}
}
\description{
  Schedules the current reactive context to be invalidated
  in the given number of milliseconds.
}

