% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.reactivevalues}
\alias{is.reactivevalues}
\title{Checks whether an object is a reactivevalues object}
\usage{
is.reactivevalues(x)
}
\arguments{
\item{x}{The object to test.}
}
\description{
Checks whether its argument is a reactivevalues object.
}
\seealso{
\code{\link{reactiveValues}}.
}

