% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{isRunning}
\alias{isRunning}
\title{Check whether a Shiny application is running}
\usage{
isRunning()
}
\value{
\code{TRUE} if a Shiny application is currently running. Otherwise,
\code{FALSE}.
}
\description{
This function tests whether a Shiny application is currently running.
}
