% Generated by roxygen2 (4.0.1): do not edit by hand
\name{knitr_methods}
\alias{knit_print.shiny.appobj}
\alias{knit_print.shiny.render.function}
\alias{knitr_methods}
\title{Knitr S3 methods}
\usage{
knit_print.shiny.appobj(x, ...)

knit_print.shiny.render.function(x, ..., inline = FALSE)
}
\arguments{
\item{x}{Object to knit_print}

\item{...}{Additional knit_print arguments}

\item{inline}{Whether the object is printed inline.}
}
\description{
These S3 methods are necessary to help Shiny applications and UI chunks embed
themselves in knitr/rmarkdown documents.
}

