% Generated by roxygen2 (4.0.1): do not edit by hand
\name{markRenderFunction}
\alias{markRenderFunction}
\title{Mark a function as a render function}
\usage{
markRenderFunction(uiFunc, renderFunc)
}
\arguments{
\item{uiFunc}{A function that renders Shiny UI. Must take a single argument:
an output ID.}

\item{renderFunc}{A function that is suitable for assigning to a Shiny output
slot.}
}
\value{
The \code{renderFunc} function, with annotations.
}
\description{
Should be called by implementers of \code{renderXXX} functions in order to
mark their return values as Shiny render functions, and to provide a hint to
Shiny regarding what UI function is most commonly used with this type of
render function. This can be used in R Markdown documents to create complete
output widgets out of just the render function.
}

