% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinywrappers.R
\name{markRenderFunction}
\alias{markRenderFunction}
\title{Mark a function as a render function}
\usage{
markRenderFunction(uiFunc, renderFunc, outputArgs = list())
}
\arguments{
\item{uiFunc}{A function that renders Shiny UI. Must take a single argument:
an output ID.}

\item{renderFunc}{A function that is suitable for assigning to a Shiny output
slot.}

\item{outputArgs}{A list of arguments to pass to the \code{uiFunc}. Render
functions should include \code{outputArgs = list()} in their own parameter
list, and pass through the value to \code{markRenderFunction}, to allow
app authors to customize outputs. (Currently, this is only supported for
dynamically generated UIs, such as those created by Shiny code snippets
embedded in R Markdown documents).}
}
\value{
The \code{renderFunc} function, with annotations.
}
\description{
Should be called by implementers of \code{renderXXX} functions in order to
mark their return values as Shiny render functions, and to provide a hint to
Shiny regarding what UI function is most commonly used with this type of
render function. This can be used in R Markdown documents to create complete
output widgets out of just the render function.
}
