% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock-session.R
\docType{data}
\name{MockShinySession}
\alias{MockShinySession}
\title{Mock Shiny Session}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MockShinySession
}
\description{
Mock Shiny Session
}
\keyword{datasets}
