% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_consulta_orc.R
\docType{data}
\name{df_orc}
\alias{df_orc}
\alias{mod_consulta_orc_server}
\title{mod_consulta_orc_ui and mod_consulta_orc_server}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 97423 rows and 39 columns.}
\usage{
df_orc

mod_consulta_orc_server(input, output, session)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{id}{shiny id}
}
\description{
A shiny Module.
}
\keyword{internal}
