% Generated by roxygen2 (4.0.1): do not edit by hand
\name{numericInput}
\alias{numericInput}
\title{Create a numeric input control}
\usage{
numericInput(inputId, label, value, min = NA, max = NA, step = NA)
}
\arguments{
\item{inputId}{Input variable to assign the control's value to}

\item{label}{Display label for the control}

\item{value}{Initial value}

\item{min}{Minimum allowed value}

\item{max}{Maximum allowed value}

\item{step}{Interval to use when stepping between min and max}
}
\value{
A numeric input control that can be added to a UI definition.
}
\description{
Create an input control for entry of numeric values
}
\examples{
numericInput("obs", "Observations:", 10,
             min = 1, max = 100)
}
\seealso{
\code{\link{updateNumericInput}}

Other input.elements: \code{\link{actionButton}},
  \code{\link{actionLink}}; \code{\link{animationOptions}},
  \code{\link{sliderInput}};
  \code{\link{checkboxGroupInput}};
  \code{\link{checkboxInput}}; \code{\link{dateInput}};
  \code{\link{dateRangeInput}}; \code{\link{fileInput}};
  \code{\link{radioButtons}}; \code{\link{selectInput}},
  \code{\link{selectizeInput}}; \code{\link{submitButton}};
  \code{\link{textInput}}
}

