% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{onStop}
\alias{onStop}
\title{Run code after app has finished}
\usage{
onStop(fun)
}
\arguments{
\item{fun}{A function that will be called after the app has finished running.}
}
\description{
This function allows you to register a callback function that is invoked
after your app has finished running (all sessions have been disconnected).
}
\examples{
onStop(function() { print("goodbye") })
}
