% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-deprecated.R
\name{pageWithSidebar}
\alias{pageWithSidebar}
\title{Create a page with a sidebar}
\usage{
pageWithSidebar(headerPanel, sidebarPanel, mainPanel)
}
\arguments{
\item{headerPanel}{The \link{headerPanel} with the application title}

\item{sidebarPanel}{The \link{sidebarPanel} containing input controls}

\item{mainPanel}{The \link{mainPanel} containing outputs}
}
\value{
A UI defintion that can be passed to the \link{shinyUI} function
}
\description{
\strong{DEPRECATED}: use \code{\link[=fluidPage]{fluidPage()}} and \code{\link[=sidebarLayout]{sidebarLayout()}} instead.
}
\keyword{internal}
