\name{panelWithSidebar}
\alias{panelWithSidebar}
\title{Create a panel that includes a sidebar for input controls
and a main panel for output}
\usage{
  panelWithSidebar(sidebarPanel, mainPanel)
}
\description{
  Create a panel that includes a sidebar for input controls
  and a main panel for output
}

