% Generated by roxygen2 (4.0.1): do not edit by hand
\name{reactivePlot}
\alias{reactivePlot}
\title{Plot output (deprecated)}
\usage{
reactivePlot(func, width = "auto", height = "auto", ...)
}
\arguments{
\item{func}{A function.}

\item{width}{Width.}

\item{height}{Height.}

\item{...}{Other arguments to pass on.}
}
\description{
See \code{\link{renderPlot}}.
}

