% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinywrappers.R
\name{reactiveTable}
\alias{reactiveTable}
\title{Table output (deprecated)}
\usage{
reactiveTable(func, ...)
}
\arguments{
\item{func}{A function.}

\item{...}{Other arguments to pass on.}
}
\description{
See \code{\link{renderTable}}.
}

