\name{runApp}
\alias{runApp}
\title{Run an application that was created by \code{\link{startApp}}. This
function does not normally return.}
\usage{
  runApp(ws_env)
}
\arguments{
  \item{ws_env}{The return value from
  \code{\link{startApp}}.}
}
\description{
  Run an application that was created by
  \code{\link{startApp}}. This function does not normally
  return.
}

