% Generated by roxygen2 (4.0.0): do not edit by hand
\name{selectInput}
\alias{selectInput}
\alias{selectizeInput}
\title{Create a select list input control}
\usage{
selectInput(inputId, label, choices, selected = NULL, multiple = FALSE)

selectizeInput(inputId, ..., options = NULL)
}
\arguments{
  \item{inputId}{Input variable to assign the control's
  value to}

  \item{label}{Display label for the control}

  \item{choices}{List of values to select from. If elements
  of the list are named then that name rather than the
  value is displayed to the user.}

  \item{selected}{The initially selected value (or multiple
  values if \code{multiple = TRUE}). If not specified then
  defaults to the first value for single-select lists and
  no values for multiple select lists.}

  \item{multiple}{Is selection of multiple items allowed?}

  \item{...}{arguments passed to \code{selectInput()}}

  \item{options}{a list of options; see the documentation
  of \pkg{selectize.js} for possible options (character
  option values inside \code{\link{I}()} will be treated as
  literal JavaScript code; see
  \code{\link{renderDataTable}()} for details)}
}
\value{
A select list control that can be added to a UI definition.
}
\description{
Create a select list that can be used to choose a single or
multiple items from a list of values.
}
\details{
\code{selectizeInput()} uses the JavaScript library \pkg{selectize.js}
(\url{https://github.com/brianreavis/selectize.js}) to extend the basic
select input element.
}
\examples{
selectInput("variable", "Variable:",
            c("Cylinders" = "cyl",
              "Transmission" = "am",
              "Gears" = "gear"))
}
\seealso{
\code{\link{updateSelectInput}}

Other input.elements: \code{\link{actionButton}};
  \code{\link{animationOptions}},
  \code{\link{sliderInput}};
  \code{\link{checkboxGroupInput}};
  \code{\link{checkboxInput}}; \code{\link{dateInput}};
  \code{\link{dateRangeInput}}; \code{\link{fileInput}};
  \code{\link{numericInput}}; \code{\link{radioButtons}};
  \code{\link{submitButton}}; \code{\link{textInput}}
}

