% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmark-state.R
\name{setBookmarkExclude}
\alias{setBookmarkExclude}
\title{Exclude inputs from bookmarking}
\usage{
setBookmarkExclude(names = character(0), session = getDefaultReactiveDomain())
}
\arguments{
\item{names}{A character vector containing names of inputs to exclude from
bookmarking.}

\item{session}{A shiny session object.}
}
\description{
This function tells Shiny which inputs should be excluded from bookmarking.
It should be called from inside the application's server function.
}
\details{
This function can also be called from a module's server function, in which
case it will exclude inputs with the specified names, from that module. It
will not affect inputs from other modules or from the top level of the Shiny
application.
}
\seealso{
\code{\link[=enableBookmarking]{enableBookmarking()}} for examples.
}
