% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\docType{package}
\name{shiny-package}
\alias{shiny-package}
\alias{shiny}
\title{Web Application Framework for R}
\description{
Shiny makes it incredibly easy to build interactive web applications with R.
Automatic "reactive" binding between inputs and outputs and extensive
prebuilt widgets make it possible to build beautiful, responsive, and
powerful applications with minimal effort.
}
\details{
The Shiny tutorial at \url{https://shiny.posit.co/tutorial/} explains
the framework in depth, walks you through building a simple application, and
includes extensive annotated examples.
}
\seealso{
\link{shiny-options} for documentation about global options.
}
\author{
\strong{Maintainer}: Carson Sievert \email{carson@posit.co} (\href{https://orcid.org/0000-0002-4958-2844}{ORCID})

Authors:
\itemize{
  \item Carson Sievert \email{carson@posit.co} (\href{https://orcid.org/0000-0002-4958-2844}{ORCID})
  \item Winston Chang \email{winston@posit.co} (\href{https://orcid.org/0000-0002-1576-2126}{ORCID})
  \item Joe Cheng \email{joe@posit.co}
  \item JJ Allaire \email{jj@posit.co}
  \item Barret Schloerke \email{barret@posit.co} (\href{https://orcid.org/0000-0001-9986-114X}{ORCID})
  \item Garrick Aden-Buie \email{garrick@adenbuie.com} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID})
  \item Yihui Xie \email{yihui@posit.co}
  \item Jeff Allen
  \item Jonathan McPherson \email{jonathan@posit.co}
  \item Alan Dipert
  \item Barbara Borges
}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
  \item  jQuery Foundation (jQuery library and jQuery UI library) [copyright holder]
  \item  jQuery contributors (jQuery library; authors listed in inst/www/shared/jquery-AUTHORS.txt) [contributor, copyright holder]
  \item  jQuery UI contributors (jQuery UI library; authors listed in inst/www/shared/jqueryui/AUTHORS.txt) [contributor, copyright holder]
  \item Mark Otto (Bootstrap library) [contributor]
  \item Jacob Thornton (Bootstrap library) [contributor]
  \item  Bootstrap contributors (Bootstrap library) [contributor]
  \item  Twitter, Inc (Bootstrap library) [copyright holder]
  \item Prem Nawaz Khan (Bootstrap accessibility plugin) [contributor]
  \item Victor Tsaran (Bootstrap accessibility plugin) [contributor]
  \item Dennis Lembree (Bootstrap accessibility plugin) [contributor]
  \item Srinivasu Chakravarthula (Bootstrap accessibility plugin) [contributor]
  \item Cathy O'Connor (Bootstrap accessibility plugin) [contributor]
  \item  PayPal, Inc (Bootstrap accessibility plugin) [copyright holder]
  \item Stefan Petre (Bootstrap-datepicker library) [contributor, copyright holder]
  \item Andrew Rowls (Bootstrap-datepicker library) [contributor, copyright holder]
  \item Brian Reavis (selectize.js library) [contributor, copyright holder]
  \item Salmen Bejaoui (selectize-plugin-a11y library) [contributor, copyright holder]
  \item Denis Ineshin (ion.rangeSlider library) [contributor, copyright holder]
  \item Sami Samhuri (Javascript strftime library) [contributor, copyright holder]
  \item  SpryMedia Limited (DataTables library) [contributor, copyright holder]
  \item Ivan Sagalaev (highlight.js library) [contributor, copyright holder]
  \item  Google LLC (lit library) [contributor, copyright holder]
  \item R Core Team (tar implementation from R) [contributor, copyright holder]
}

}
