\docType{package}
\name{shiny-package}
\alias{shiny}
\alias{shiny-package}
\title{Web Application Framework for R}
\description{
  Shiny makes it incredibly easy to build interactive web
  applications with R. Automatic "reactive" binding between
  inputs and outputs and extensive pre-built widgets make
  it possible to build beautiful, responsive, and powerful
  applications with minimal effort.
}
\details{
  The Shiny tutorial at
  \url{http://rstudio.github.com/shiny/tutorial} explains
  the framework in depth, walks you through building a
  simple application, and includes extensive annotated
  examples.
}

