\name{shiny-package}
\alias{shiny-package}
\alias{shiny}
\docType{package}
\title{
Web Application Framework for R
}
\description{
Shiny makes it incredibly easy to build interactive web 
    applications with R. Automatic "reactive" binding between inputs and 
    outputs and extensive pre-built widgets make it possible to build 
    beautiful, responsive, and powerful applications with minimal effort.
    
    The Shiny tutorial at \url{http://rstudio.github.com/shiny/tutorial}
    explains the framework in-depth, walks you through
    building a simple application, and includes extensive annotated 
    examples.
}
\details{
\tabular{ll}{
Package: \tab shiny\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2012-07-28\cr
License: \tab GPL-3\cr
Depends: \tab R (>= 2.14.1), methods, websockets (>= 1.1.4), caTools, RJSONIO, xtable\cr
Imports: \tab stats, tools, utils, datasets\cr
URL: \tab https://github.com/rstudio/shiny, http://rstudio.github.com/shiny/tutorial\cr
BugReports: \tab https://github.com/rstudio/shiny/issues\cr
}

}
\author{
RStudio, Inc.

Maintainer: Joe Cheng <joe@rstudio.org>
}


\keyword{ package }

