% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\docType{package}
\name{shiny-package}
\alias{shiny-package}
\alias{shiny}
\title{Web Application Framework for R}
\description{
Shiny makes it incredibly easy to build interactive web applications with R.
Automatic "reactive" binding between inputs and outputs and extensive
prebuilt widgets make it possible to build beautiful, responsive, and
powerful applications with minimal effort.
}
\details{
The Shiny tutorial at \url{http://shiny.rstudio.com/tutorial/} explains
the framework in depth, walks you through building a simple application, and
includes extensive annotated examples.
}
\seealso{
\link{shiny-options} for documentation about global options.
}
