% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_template.R
\name{shinyAppTemplate}
\alias{shinyAppTemplate}
\title{Generate a Shiny application from a template}
\usage{
shinyAppTemplate(path = NULL, examples = "default")
}
\arguments{
\item{path}{Path to create new shiny application template.}

\item{examples}{Either one of "default", "ask", "all", or any combination of
"app", "rdir", "module", "shinytest", "testthat", and "server". In an
interactive session, "default" falls back to "ask"; in a non-interactive
session, "default" falls back to "all". With "ask", this function will
prompt the user to select which template items will be added to the new app
directory. With "all", all template items will be added to the app
directory.}
}
\description{
This function populates a directory with files for a Shiny application. They
are based off of the "12_counter" example which can be run with
\code{runExample()}.
}
\details{
In an interactive R session, this function will, by default, prompt the user
which components to add to the application.

The full example application includes the following files and directories:\preformatted{appdir/
├── app.R
├── R
│   ├── my-module.R
│   └── utils.R
└── tests
    ├── server.R
    ├── server
    │   ├── test-mymodule.R
    │   └── test-server.R
    ├── shinytest.R
    ├── shinytest
    │   └── mytest.R
    ├── testthat.R
    └── testthat
        ├── helper-load.R
        └── test-utils.R
}

Some notes about these files:
\itemize{
\item app.R is the main application file.
\item All files in the R/ subdirectory are automatically sourced when the
application is run.
\item The R/my-module.R file is automatically sourced when the application
is run. This file contains code for a \href{moduleServer()}{Shiny module} which
is used in the application.
\item The tests/ directory contains various tests for the application. You may
choose to use or remove any of them. They can be executed by the
\code{\link[=runTests]{runTests()}} function.
\item tests/server.R is a test runner for test files in
tests/server/.
\item tests/server/test-mymodule.R is a test for the module.
\item tests/shinytest.R is a test runner for test files in the
tests/shinytest/ directory.
\item tests/shinytest/mytest.R is a test that uses the
\href{https://rstudio.github.io/shinytest/}{shinytest} package to do
snapshot-based testing.
\item tests/testthat.R is a test runner for test files in the
tests/testthat/ directory.
\item tests/testthat/helper-load.R is a helper script that is automatically
loaded before running test-counter.R. (This is performed by the testthat
package.)
\item tests/testthat/test-utils.R is a set of tests that use the
\href{https://testthat.r-lib.org/}{testthat} package for testing.
}
}
