% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_template.R
\name{shinyAppTemplate}
\alias{shinyAppTemplate}
\title{Generate a Shiny application template}
\usage{
shinyAppTemplate(
  path = NULL,
  ...,
  examples = TRUE,
  quiet = FALSE,
  interactive = base::interactive()
)
}
\arguments{
\item{path}{Path to create new shiny application template.}

\item{examples}{Should the new path include example code? Defaults to \code{TRUE}.}

\item{quiet}{Should status information be printed? Defaults to \code{FALSE}.}

\item{interactive}{If \code{TRUE}, the user may be asked questions and prompted
for a response.}
}
\description{
This function populates a directory with files for a Shiny application. They
are based off of the "12_counter" example which can be run with
\code{runExample()}.
}
\details{
The example application includes the following files and directories:\preformatted{appdir/
├── app.R
├── R
│   └── counter-module.R
└── tests
    ├── integration.R
    ├── integration
    │   └── test-counter.R
    ├── shinytest.R
    ├── shinytests
    │   └── mytest.R
    ├── testthat.R
    └── testthat
        ├── helper-load.R
        └── test-counter.R
}

Some notes about these files:
\itemize{
\item app.R is the main application file.
\item All files in the R/ subdirectory are automatically sourced when the
application is run.
\item The R/counter-module.R file is automatically sourced when the application
is run. This file contains code for a \href{moduleServer()}{Shiny module} which
is used in the application.
\item The tests/ directory contains various tests for the application. You may
choose to use or remove any of them. They can be executed by the
\code{\link[=runTests]{runTests()}} function.
\item tests/integration.R is a test runner for test files in the
tests/integration/ directory.
\item tests/integration/test-counter.R is a test for the counter module.
\item tests/shinytest.R is a test runner for test files in the
tests/shinytest/ directory.
\item tests/shinytest/mytest.R is a test that uses the
\href{https://rstudio.github.io/shinytest/}{shinytest} package to do
snapshot-based testing.
\item tests/testthat.R is a test runner for test files in the
tests/testthat/ directory.
\item tests/testthat/test-counter.R is a set of tests that use the
\href{https://testthat.r-lib.org/}{testthat} package for testing.
\item tests/testthat/helper-load.R is a helper script that is automatically
loaded before running test-counter.R. (This is performed by the testthat
package.)
}
}
