% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{shinyAppTemplate}
\alias{shinyAppTemplate}
\title{Generate a Shiny application template}
\usage{
shinyAppTemplate(
  path = NULL,
  ...,
  examples = TRUE,
  quiet = FALSE,
  interactive = interactive()
)
}
\arguments{
\item{path}{Path to create new shiny application template.}

\item{examples}{Should the new path include example code? Defaults to \code{TRUE}.}

\item{quiet}{Should status information be printed? Defaults to \code{FALSE}.}

\item{interactive}{If \code{TRUE}, the user may be asked questions and prompted
for a response.}
}
\description{
Generate a Shiny application template
}
