% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shinyDeprecated}
\alias{shinyDeprecated}
\title{Print message for deprecated functions in Shiny}
\usage{
shinyDeprecated(
  new = NULL,
  msg = NULL,
  old = as.character(sys.call(sys.parent()))[1L],
  version = NULL
)
}
\arguments{
\item{new}{Name of replacement function.}

\item{msg}{Message to print. If used, this will override the default message.}

\item{old}{Name of deprecated function.}

\item{version}{The last version of Shiny before the item was deprecated.}
}
\description{
To disable these messages, use \code{options(shiny.deprecation.messages=FALSE)}.
}
\keyword{internal}
