% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{showModal}
\alias{showModal}
\alias{removeModal}
\title{Show or remove a modal dialog}
\usage{
showModal(ui, session = getDefaultReactiveDomain())

removeModal(session = getDefaultReactiveDomain())
}
\arguments{
\item{ui}{UI content to show in the modal.}

\item{session}{The \code{session} object passed to function given to
\code{shinyServer}.}
}
\description{
This causes a modal dialog to be displayed in the client browser, and is
typically used with \code{\link[=modalDialog]{modalDialog()}}.
}
\seealso{
\code{\link[=modalDialog]{modalDialog()}} for examples.
}
