% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-ui.R
\name{showTab}
\alias{showTab}
\title{Dynamically show a tabPanel}
\usage{
showTab(tabsetPanelId, target, immediate = FALSE,
  session = getDefaultReactiveDomain())
}
\arguments{
\item{tabsetPanelId}{The \code{id} of the \code{tabsetPanel()} that
\code{target} belongs to.}

\item{target}{The \code{value} of the \code{tabPanel()} to be shown.}

\item{immediate}{Whether \code{tab} should be immediately shown from when
you call \code{removeTab}, or whether Shiny should wait until all
outputs have been updated and all observers have been run (default).}

\item{session}{The shiny session within which to call \code{showTab}.}
}
\description{
Dynamically show (expose) a hidden \code{tabPanel()} from an existing
\code{tabsetPanel}.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
ui <- fluidPage(
  sidebarLayout(
    sidebarPanel(actionButton("show", "Show tab")),
    mainPanel(
      tabsetPanel(id = "tabs",
        tabPanel("Hello", "This is the hello tab"),
        tabPanel("Foo", "This is the foo tab"),
        tabPanel("Bar", "This is the bar tab")
      )
    )
  )
)
server <- function(input, output, session) {
  # Hide tab as soon as app starts up
  hideTab(tabsetPanelId = "tabs", target = "Foo")

  observeEvent(input$show, {
    showTab(tabsetPanelId = "tabs", target = "Foo")
  })
}

shinyApp(ui, server)
}
}
\seealso{
\code{\link{insertTab}}, \code{\link{removeTab}},
  \code{\link{hideTab}}
}
