% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-layout.R
\name{sidebarLayout}
\alias{sidebarLayout}
\title{Layout a sidebar and main area}
\usage{
sidebarLayout(sidebarPanel, mainPanel, position = c("left", "right"),
  fluid = TRUE)
}
\arguments{
\item{sidebarPanel}{The \link{sidebarPanel} containing input controls}

\item{mainPanel}{The \link{mainPanel} containing outputs}

\item{position}{The position of the sidebar relative to the main area ("left"
or "right")}

\item{fluid}{\code{TRUE} to use fluid layout; \code{FALSE} to use fixed
layout.}
}
\description{
Create a layout with a sidebar and main area. The sidebar is displayed with a
distinct background color and typically contains input controls. The main
area occupies 2/3 of the horizontal width and typically contains outputs.
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {
options(device.ask.default = FALSE)

# Define UI
ui <- fluidPage(

  # Application title
  titlePanel("Hello Shiny!"),

  sidebarLayout(

    # Sidebar with a slider input
    sidebarPanel(
      sliderInput("obs",
                  "Number of observations:",
                  min = 0,
                  max = 1000,
                  value = 500)
    ),

    # Show a plot of the generated distribution
    mainPanel(
      plotOutput("distPlot")
    )
  )
)

# Server logic
server <- function(input, output) {
  output$distPlot <- renderPlot({
    hist(rnorm(input$obs))
  })
}

# Complete app with UI and server components
shinyApp(ui, server)
}
}
\seealso{
Other layout functions: \code{\link{fillRow}},
  \code{\link{fixedPage}}, \code{\link{flowLayout}},
  \code{\link{fluidPage}}, \code{\link{splitLayout}},
  \code{\link{titlePanel}}, \code{\link{verticalLayout}}
}
\concept{layout functions}
