% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{stopApp}
\alias{stopApp}
\title{Stop the currently running Shiny app}
\usage{
stopApp(returnValue = invisible())
}
\arguments{
\item{returnValue}{The value that should be returned from
\code{\link[=runApp]{runApp()}}.}
}
\description{
Stops the currently running Shiny app, returning control to the caller of
\code{\link[=runApp]{runApp()}}.
}
