% Generated by roxygen2 (4.0.1): do not edit by hand
\name{submitButton}
\alias{submitButton}
\title{Create a submit button}
\usage{
submitButton(text = "Apply Changes", icon = NULL)
}
\arguments{
\item{text}{Button caption}

\item{icon}{Optional \code{\link{icon}} to appear on the button}
}
\value{
A submit button that can be added to a UI definition.
}
\description{
Create a submit button for an input form. Forms that include a submit
button do not automatically update their outputs when inputs change,
rather they wait until the user explicitly clicks the submit button.
}
\examples{
submitButton("Update View")
submitButton("Update View", icon("refresh"))
}
\seealso{
Other input.elements: \code{\link{actionButton}},
  \code{\link{actionLink}}; \code{\link{animationOptions}},
  \code{\link{sliderInput}};
  \code{\link{checkboxGroupInput}};
  \code{\link{checkboxInput}}; \code{\link{dateInput}};
  \code{\link{dateRangeInput}}; \code{\link{fileInput}};
  \code{\link{numericInput}}; \code{\link{radioButtons}};
  \code{\link{selectInput}}, \code{\link{selectizeInput}};
  \code{\link{textInput}}
}

