% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{tableOutput}
\alias{tableOutput}
\alias{dataTableOutput}
\title{Create a table output element}
\usage{
tableOutput(outputId, class = NULL, ...)

dataTableOutput(outputId, class = NULL, ...)
}
\arguments{
\item{outputId}{output variable to read the table from}

\item{class}{Additional CSS classes to apply to the div tag, if any.}

\item{...}{Additional arguments to apply to the div tag, if any.}
}
\value{
A table output element that can be included in a panel
}
\description{
Render a \code{\link[=renderTable]{renderTable()}} or \code{\link[=renderDataTable]{renderDataTable()}} within an
application page. \code{renderTable} uses a standard HTML table, while
\code{renderDataTable} uses the DataTables Javascript library to create an
interactive table with more features.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
  # table example
  shinyApp(
    ui = fluidPage(
      fluidRow(
        column(12,
          tableOutput('table')
        )
      )
    ),
    server = function(input, output) {
      output$table <- renderTable(iris)
    }
  )


  # DataTables example
  shinyApp(
    ui = fluidPage(
      fluidRow(
        column(12,
          dataTableOutput('table')
        )
      )
    ),
    server = function(input, output) {
      output$table <- renderDataTable(iris)
    }
  )
}
}
\seealso{
\code{\link[=renderTable]{renderTable()}}, \code{\link[=renderDataTable]{renderDataTable()}}.
}
