% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{tabPanel}
\alias{tabPanel}
\alias{tabPanelBody}
\alias{navItem}
\alias{navSpacer}
\title{Create navigation items}
\usage{
tabPanel(title, ..., value = title, icon = NULL)

tabPanelBody(value, ..., icon = NULL)

navItem(...)

navSpacer()
}
\arguments{
\item{title}{\verb{[character(1)]}\cr A display title for the nav item.}

\item{...}{\verb{[UI elements]}\cr
\itemize{
\item For \code{tabPanel()}/\code{tabPanelBody()}: content to display when the tab is
active.
\item For \code{navLink()}: arguments to the \code{\link[=a]{a()}} tag.
\item For \code{navForm()}: the input form to place inline in the nav container.
}}

\item{value}{\verb{[character(1)]}\cr A value associate with a \code{tabPanel()} to
determine when it is selected. This value may be supplied to the relevant
nav container's \code{selected} argument to show the tab immediately on page
load. This value is also useful for programmatically updating the selected
tab via \code{\link[=updateTabsetPanel]{updateTabsetPanel()}}, \code{\link[=showTab]{showTab()}}, etc (updating selected tabs
this way is often useful for showing/hiding panels of content via other UI
controls like \code{\link[=radioButtons]{radioButtons()}} -- in this scenario, consider using
\code{\link[=tabPanelBody]{tabPanelBody()}} with \code{tabsetPanel(type = "hidden")}).}

\item{icon}{An \code{\link[=icon]{icon()}} to appear to the left of the nav item's \code{title}.}
}
\value{
A nav item that may be passed to a nav container (e.g. \code{\link[=tabsetPanel]{tabsetPanel()}}
or \code{\link[=navbarPage]{navbarPage()}}).
}
\description{
Create "nav items" for use inside "nav containers" (e.g., \code{\link[=tabsetPanel]{tabsetPanel()}} or
\code{\link[=navbarPage]{navbarPage()}}). Individual items are created with one of either \code{tabPanel()}
(navigate to a panel of content), \code{navLink()} (navigate to remote content),
or \code{navForm()} (for custom inline navigation, like search forms). Any
collection of nav items can be grouped together in a menu with
\code{navbarMenu()}.
}
\examples{

panels <- list(
  tabPanel("Plot", "plot"),
  tabPanel("Summary", "summary"),
  tabPanel("Table", "table"),
  navbarMenu(
    "External links",
    navLink("Shiny", "https://github.com/rstudio/shiny", icon("github"))
  )
)

if (interactive()) {
  shinyApp(
    fluidPage(tabsetPanel(!!!panels, card = TRUE)),
    function(...) {}
  )
  shinyApp(
    navbarPage("Title", !!!panels[1:3], right = panels[[4]]),
    function(...) {}
  )
}

}
\seealso{
\code{\link[=tabsetPanel]{tabsetPanel()}}, \code{\link[=navbarPage]{navbarPage()}}, \code{\link[=navlistPanel]{navlistPanel()}}
}
