% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{testApp}
\alias{testApp}
\title{Runs the tests associated with this Shiny app}
\usage{
testApp(appDir = ".")
}
\arguments{
\item{appDir}{The base directory for the application.}
}
\description{
Sources the \code{.R} files in the top-level of \code{tests/} much like \code{R CMD check}.
These files are typically simple runners for tests nested in other
directories under \code{tests/}.
}
\details{
TODO: add in filters for specific tests, or high-level toggles for particular types of tests. Or should that just be a filter on the runners?

Historically, \href{https://rstudio.github.io/shinytest/}{shinytest}
recommended placing tests at the top-level of the \code{tests/} directory. In
order to support that model, \code{testApp} first checks to see if the \code{.R}
files in the \code{tests/} directory are all shinytests; if so, just calls out
to \code{\link[shinytest:testApp]{shinytest::testApp()}}.
}
