% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-module.R
\name{testServer}
\alias{testServer}
\title{Test an app's server-side logic}
\usage{
testServer(expr, dir = NULL)
}
\arguments{
\item{expr}{Test code containing expectations}

\item{dir}{The directory root of the Shiny application. If \code{NULL}, this function
will work up the directory hierarchy --- starting with the current directory ---
looking for a directory that contains an \code{app.R} or \code{server.R} file.}
}
\description{
Test an app's server-side logic
}
