% Generated by roxygen2 (4.0.1): do not edit by hand
\name{titlePanel}
\alias{titlePanel}
\title{Create a panel containing an application title.}
\usage{
titlePanel(title, windowTitle = title)
}
\arguments{
\item{title}{An application title to display}

\item{windowTitle}{The title that should be displayed by the browser window.}
}
\description{
Create a panel containing an application title.
}
\details{
Calling this function has the side effect of including a
  \code{title} tag within the head. You can also specify a page title
  explicitly using the `title` parameter of the top-level page function.
}
\examples{
titlePanel("Hello Shiny!")
}

