% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-layout.R
\name{titlePanel}
\alias{titlePanel}
\title{Create a panel containing an application title.}
\usage{
titlePanel(title, windowTitle = title)
}
\arguments{
\item{title}{An application title to display}

\item{windowTitle}{The title that should be displayed by the browser window.}
}
\description{
Create a panel containing an application title.
}
\details{
Calling this function has the side effect of including a
\code{title} tag within the head. You can also specify a page title
explicitly using the \code{title} parameter of the top-level page function.
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {

ui <- fluidPage(
  titlePanel("Hello Shiny!")
)
shinyApp(ui, server = function(input, output) { })
}
}
