% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wellPanel}
\alias{wellPanel}
\title{Create a well panel}
\usage{
wellPanel(...)
}
\arguments{
\item{...}{UI elements to include inside the panel.}
}
\value{
The newly created panel.
}
\description{
Creates a panel with a slightly inset border and grey background. Equivalent
to Bootstrap's \code{well} CSS class.
}

