% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{wellPanel}
\alias{wellPanel}
\title{Create a well panel}
\usage{
wellPanel(...)
}
\arguments{
\item{...}{UI elements to include inside the panel.}
}
\value{
The newly created panel.
}
\description{
Creates a panel with a slightly inset border and grey background. Equivalent
to Bootstrap's \code{well} CSS class.
}
